+++ docs/source/references.bib
@@ -1,766 +1,778 @@
-@article{Cavendish1974,
-abstract = {The object of this paper is to describe a new algorithm for the semi‐automatic triangulation of arbitrary, multiply connected planar domains. The strategy is based upon a modification of a finite element mesh genration algorithm recently developed. 1 The scheme is designed for maximum flexibility and is capable of generating meshes of triangular elements for the decomposition of virtually any multiply connected planar domain. Moreover, the desired density of elements in various regions of the problem domain is specified by the user, thus allowing him to obtain a mesh decomposition appropriate to the physical loading and/or boundary conditions of the particular problem at hand. Several examples are presented to illustrate the applicability of the algorithm. An extension of the algorithm to the triangulation of shell structures is indicated. Copyright {\textcopyright} 1974 John Wiley {\&} Sons, Ltd},
-author = {Cavendish, James C.},
-doi = {10.1002/nme.1620080402},
-file = {:home/tim/Nextcloud/Documents/Papers/1974 - Cavendish - Automatic triangulation of arbitrary planar domains for the finite element method.pdf:pdf},
-issn = {10970207},
-journal = {International Journal for Numerical Methods in Engineering},
-number = {4},
-pages = {679--696},
-title = {{Automatic triangulation of arbitrary planar domains for the finite element method}},
-volume = {8},
-year = {1974}
+@article{Dassi2018,
+   abstract = {Given a tetrahedral mesh and objective functionals measuring the mesh quality which take into account the shape, size, and orientation of the mesh elements, our aim is to improve the mesh quality as much as possible. In this paper, we combine the moving mesh smoothing, based on the integration of an ordinary differential equation coming from a given functional, with the lazy flip technique, a reversible edge removal algorithm to modify the mesh connectivity. Moreover, we utilize radial basis function (RBF) surface reconstruction to improve tetrahedral meshes with curved boundary surfaces. Numerical tests show that the combination of these techniques into a mesh improvement framework achieves results which are comparable and even better than the previously reported ones.},
+   author = {Franco Dassi and Lennard Kamenski and Patricio Farrell and Hang Si},
+   doi = {10.1016/j.cad.2017.11.010},
+   issn = {00104485},
+   journal = {CAD Computer Aided Design},
+   keywords = {Edge flipping,Mesh improvement,Mesh quality,Mesh smoothing,Moving mesh,Radial basis functions},
+   pages = {2-13},
+   title = {Tetrahedral mesh improvement using moving mesh smoothing, lazy searching flips, and RBF surface reconstruction},
+   volume = {103},
+   url = {https://doi.org/10.1016/j.cad.2017.11.010},
+   year = {2018}
 }
-@article{Carr2001,
[…]
