+++ _sources/examples/MeshModification/sg_execution_times.rst.txt
@@ -6,7 +6,7 @@
 
 Computation times
 =================
-**01:29.476** total execution time for 2 files **from examples/MeshModification**:
+**01:39.018** total execution time for 2 files **from examples/MeshModification**:
 
 .. container::
 
@@ -33,8 +33,8 @@ Computation times
      - Time
      - Mem (MB)
    * - :ref:`sphx_glr_examples_MeshModification_demo_registration.py` (``demo_registration.py``)
-     - 01:13.694
+     - 01:22.251
      - 0.0
    * - :ref:`sphx_glr_examples_MeshModification_demo_smoothing.py` (``demo_smoothing.py``)
-     - 00:15.782
+     - 00:16.767
      - 0.0
