+++ _sources/examples/MeshModification/demo_smoothing.rst.txt
@@ -262,7 +262,7 @@ and geometries.
 
 .. rst-class:: sphx-glr-timing
 
-   **Total running time of the script:** (0 minutes 15.782 seconds)
+   **Total running time of the script:** (0 minutes 16.767 seconds)
 
 
 .. _sphx_glr_download_examples_MeshModification_demo_smoothing.py:
