+++ _sources/examples/MeshModification/demo_registration.rst.txt
@@ -180,7 +180,7 @@ Image-to-Image Registration
 
 .. rst-class:: sphx-glr-timing
 
-   **Total running time of the script:** (1 minutes 13.694 seconds)
+   **Total running time of the script:** (1 minutes 22.251 seconds)
 
 
 .. _sphx_glr_download_examples_MeshModification_demo_registration.py:
