+++ _sources/examples/MeshGeneration/demo_mixed-top.rst.txt
@@ -137,7 +137,7 @@ including :func:`~mymesh.implicit.gyroid`, :func:`~mymesh.implicit.primitive`
 
 .. rst-class:: sphx-glr-timing
 
-   **Total running time of the script:** (0 minutes 8.393 seconds)
+   **Total running time of the script:** (0 minutes 8.162 seconds)
 
 
 .. _sphx_glr_download_examples_MeshGeneration_demo_mixed-top.py:
