+++ _sources/examples/MeshAnalysis/sg_execution_times.rst.txt
@@ -6,7 +6,7 @@
 
 Computation times
 =================
-**01:22.784** total execution time for 1 file **from examples/MeshAnalysis**:
+**01:25.896** total execution time for 1 file **from examples/MeshAnalysis**:
 
 .. container::
 
@@ -33,5 +33,5 @@ Computation times
      - Time
      - Mem (MB)
    * - :ref:`sphx_glr_examples_MeshAnalysis_demo_curvature.py` (``demo_curvature.py``)
-     - 01:22.784
+     - 01:25.896
      - 0.0
