+++ _sources/examples/MeshGeneration/demo_pasta.rst.txt
@@ -133,7 +133,7 @@ Please submit any new pasta constructions to toj@bu.edu
 
 .. rst-class:: sphx-glr-timing
 
-   **Total running time of the script:** (0 minutes 9.274 seconds)
+   **Total running time of the script:** (0 minutes 9.426 seconds)
 
 
 .. _sphx_glr_download_examples_MeshGeneration_demo_pasta.py:
