### 2025-12-28 d69530c
feat(QuadtreeNode): set up preliminary nearest neighbor search
