### 2026-01-07 #9641 @RichardLitt
@TimJosephson JOSS has just updated its requirements for papers. Can you take a gander at https://joss.readthedocs.io/en/latest/submitting.html#scope-and-significance and let me know if you think this article still applies, as a pre-check before I send it out for review?
