### 2026-01-08 55f2dd7
feat(rays): add PintsInBox, PointsInBox2D
