### 2026-01-08 e76c36c
feat(tree): add KDtree and KNN search for octree
