+++ src/mymesh/tree.py
@@ -90,7 +90,7 @@ Octree Querying
 
 """
 import numpy as np
-from . import rays, utils, mesh
+from . import rays, utils, mesh, primitives
 import sympy as sp
 import copy, heapq
 
@@ -220,7 +220,6 @@ class TreeNode:
                 node.children = []
         return pruned        
 
-
 class QuadtreeNode(TreeNode):
           
     def __init__(self,centroid,size,parent=None,data=None,level=0,state='unknown'):
@@ -349,7 +348,8 @@ class QuadtreeNode(TreeNode):
             the node, otherwise False.
         """
         limits = self.getLimits()
-        inside =  np.array([rays.PointInBox2D(point, *limits, inclusive=inclusive) for point in points])
+        # inside =  np.array([rays.PointInBox2D(point, *limits, inclusive=inclusive) for point in points])
+        inside = rays.PointsInBox2D(points, *limits, inclusive=inclusive)
         
         return inside
     
@@ -367,7 +367,9 @@ class QuadtreeNode(TreeNode):
         out : list
             List of indices of the points that are contained within the node
         """
-        out = [idx for idx,point in enumerate(points) if self.PointInNode(point)]
+        # out = [idx for idx,point in enumerate(points) if self.PointInNode(point)]
+        inside = self.PointsInNode(points)
+        out = np.where(inside) # can probably skip where and use logical indexing (which would kinda make this whole function obsolete?)
 
         return out
      
@@ -494,29 +496,48 @@ class QuadtreeNode(TreeNode):
             elif len(edgesInChild) == 0:
                 child.state = 'empty'
 
-    def search_pt(self, x):
-
-        if self.PointInNode(x):
-            if len(self.children) == 0:
-                return self
-            for child in self.children:
-                out = child.search_pt(x)
-                if out is not None:
-                    return out
-        else:
-            return None
-        
     def node_distance(self, x):
-        # closest distance between  between a node in the tree and a point
+        """
+        Calculate the nearest distance between a point and a quadtree node.
+
+        Parameters
+        ----------
+        x : np.ndarray
+            Point coordinates (shape=(2,))
+
+        Returns
+        -------
+        dist : float
+            Distance between a point and the node.
+        """        
         limits = self.getLimits()
         nearest_point = np.clip(x, limits[:, 0], limits[:, 1])
         
         dist = np.linalg.norm(np.asarray(x) - nearest_point)
         return dist
 
-
     def query_knn(self, x, k=1):
+        """
+        Find the k-nearest neighbor points in the tree to a point.
+        Note: This currently only works for a point cloud tree (e.g. created
+        with :func:`Points2Quadtree`)
 
+        Parameters
+        ----------
+        x : np.ndarray
+            Point coordinates (shape=(2,))
+        k : int, optional
+            Number of nearest neighbor points to identify, by default 1
+
+        Returns
+        -------
+        dist : np.ndarray
+            Array of distance between :code:`x` and the nearest neighbors 
+            (shape=(k,))
+        coordinates : np.ndarray
+            Array of coordinates for the nearest neighbor points
+            (shape=(k,2))
+        """        
         # Get point-root distance
         if self.PointInNode(x):
             dist = 0
@@ -545,17 +566,20 @@ class QuadtreeNode(TreeNode):
                             dist = 0
                         else:
                             dist = child.node_distance(x)
-                        heapq.heappush(queue, (dist,child))
+                        if len(queue) >= k:
+                            if dist <= queue[0][0]:
+                                heapq.heappush(queue, (dist,child))
+                        else:
+                            heapq.heappush(queue, (dist,child))
             else:
[…]
