### 2026-01-08 0dadbaf
docs: add verification section
