+++ docs/source/verify.rst
@@ -1,0 +1,9 @@
+Verification
+============
+
+Verification and performance evaluation of select operations.
+
+.. toctree::
+    :maxdepth: 2
+    
+    verification/tree
