### 2026-01-12 4a6c65e
docs(paper): update manuscript to meet new JOSS requirements
