+++ paper/paper.md
@@ -30,7 +30,8 @@ bibliography: paper.bib
 
 A mesh is a discrete representation that subdivides a geometry or computational domain into a collection of points (nodes) connected by simple shapes (elements).
 Meshes are used for a variety of purposes, including simulations (e.g. finite element, finite volume, and finite difference methods), visualization & computer graphics, image analysis, and additive manufacturing.
-`mymesh` is a general purpose set of tools for generating, manipulating, and analyzing meshes. `mymesh` is particularly focused on implicit function and image-based meshing, with other functionality including:
+`mymesh` is a general purpose set of tools for generating, manipulating, and analyzing meshes. 
+`mymesh` is particularly focused on implicit function and image-based meshing, with other functionality including:
 
 - geometric and curvature analysis,
 - intersection and inclusion tests (e.g. ray-surface intersection and point-in-surface tests)
@@ -40,10 +41,6 @@ Meshes are used for a variety of purposes, including simulations (e.g. finite el
 - mesh quality evaluation and improvement,
 - mesh type conversion (e.g. volume to surface, hexahedral or mixed-element to tetrahedral, first-order elements to second-order elements).
 
-`mymesh` was originally developed in support of research within the Skeletal Mechanobiology and Biomechanics Lab at Boston University. 
-It was used extensively in the scaffold design optimization research by @Josephson2024b and is currently being used in various ongoing projects, including vertebral modeling, hip fracture modeling, growth modeling of skeletal tissue, and analysis of objects imaged using micro-computed tomography (μCT). 
-`mymesh` has proven useful in a variety of research applications, well beyond those that inspired its original development, and we expect it to remain a valuable tool in future research efforts.  
-
 # Statement of need
 
 Mesh-based representations of geometries are essential in a wide variety of research applications, and as such, there is a need for robust, efficient, and easy-to-use software for creating, analyzing, and manipulating meshes.
@@ -52,12 +49,32 @@ Some are general purpose, like CGAL [@cgal], VTK [@Schroeder2006], and Gmsh [@Ge
 In Python, most meshing packages depend on (or are direct wrappers to) one or more of these libraries, such as PyVista [@Sullivan2019] (a pythonic interface to VTK), MeshPy (which interfaces to Triangle and TetGen), and PyMesh (which depends on CGAL, Triangle, TetGen, and others). 
 While these interfaces are useful and provide access to powerful mesh generation tools, their reliance on external dependencies can make them less easy to use and limit code readability, making it more difficult to understand how the code works. 
 TriMesh [@trimesh] stands out as a capable, pure-Python library focused on triangular surface meshes, but it isn't intended for use with quadrilateral, mixed-element, or volumetric meshes. 
-There is thus a need for a full-featured, accessible, and easy to use Python package for creating and working with meshes.
+Given the intended focus and/or design philosophies of these existing softwares, it was determined that building `mymesh`, rather than making contributions to existing software, was the best way to achieve a full-featured, accessible, and easy to use Python package for creating and working with meshes. 
 
 `mymesh` strives to meet this need as a library of meshing tools, written in Python, with clear documentation that makes it both easy to use and easy to understand.
 `mymesh` has a particular focus on implicit function and image-based meshes, but also supplies a wide variety of general purpose tools. 
 Rather than wrapping other libraries, algorithms are implemented from scratch, often based on or inspired by published algorithms and research. 
 By providing an easily usable interface to both high-level and low-level functionality, we hope to provide both complete solutions and a set of building blocks for the development of other mesh-related tools.
+
+# Research impact statement
+
[…]
