### 2026-01-12 #9641 @danielskatz
@TimJosephson - I notice some formatting (bullets) didn't quite work right. You might want to change the .md file, then regenerate the PDF using the same editorialbot command I used above.
