### 2026-01-12 542fbd5
docs(paper): fix formatting of bullets
