### 2026-01-12 b0010e1
docs(paper): fix formatting of bullets
