### 2026-01-27 #9641 @RichardLitt
@nschloe, @jorgensd: Would either of you be interested? Danke and takk.
