### 2026-01-27 d580063
feat(tree): Add KDTree query_knn
