+++ docs/source/verification/tree.rst
@@ -5,14 +5,23 @@ Tree
 K-Nearest Neighbor Search
 -------------------------
 
+Spatial tree structures can dramatically speed up nearest neighbor searching compared to brute-force approaches that require distances between all points to be calculated, however, there is an added up-front cost of constructing the tree.
+If only a single search needs to be performed, the brute-force approach may be the most efficient option, but if performing multiple searches, or other operations that can utilize the tree structure, performance can be enhanced significantly.
+
+Here the performance of multiple tree structures are compared to a brute-force search.
+A `quadtree <QuadtreeNode>`,  `KD tree <KDtreeNode>`, and SciPy's implementation of a `KD tree <https://docs.scipy.org/doc/scipy/reference/generated/scipy.spatial.KDTree.html>`_ are compared to a brute force search for 2D point data.
+For 3D point data, an `octree <OctreeNode>`, is used in place of the quadtree.
+Note that the KD tree can be used for any k-dimensional data.
+
 2D Point Data
 #############
 
 .. code::
-
-    points = np.random.rand(2,3)
-    x = np.random.rand(2)
-    k = 3
+    
+    d = 2       # Dimensions
+    k = 3       # Number of nearest neighbors
+    points = np.random.rand(n,d)
+    x = np.random.rand(n)
 
     # quadtree build
     quadtree = tree.Points2Quadtree(points)
@@ -21,11 +30,17 @@ K-Nearest Neighbor Search
     qtree_out = quadtree.query_knn(x, k=k)
 
     # kdtree build
-    kdtree = scipy.spatial.KDTree(points)
+    kdtree = tree.Points2KDtree(points)
 
     # kdtree search
-    kdtree_out = kdtree.query(x, k=k)
-    kdtree_out = (np.atleast_1d(kdtree_out[0]), np.atleast_2d(points[kdtree_out[1]]))
+    kdtree_out = kdtree.query_knn(x, k=k)
+
+    # scipy kdtree build
+    sci_kdtree = scipy.spatial.KDTree(points)
+
+    # scipy kdtree search
+    sci_kdtree_out = sci_kdtree.query(x, k=k)
+    sci_kdtree_out = (np.atleast_1d(sci_kdtree_out[0]), np.atleast_2d(points[sci_kdtree_out[1]]))
 
     # brute-force
     d = scipy.spatial.distance.cdist(points, np.atleast_2d(x)).flatten()
@@ -44,26 +59,29 @@ K-Nearest Neighbor Search
 
     npoints = np.array([1e1, 1e2, 1e3, 1e4, 1e5, 1e6], dtype=int)
 
+    sci_kdtree_build_time = []
+    sci_kdtree_search_time = []
     kdtree_build_time = []
     kdtree_search_time = []
     qtree_build_time = []
     qtree_search_time = []
     brute_search_time = []
-
-    x = np.random.rand(2)
+    
     k = 3
+    d = 2
+    x = np.random.rand(d)
 
     # pre-run for any first time pre-compiling
-    points = np.random.rand(10,2)
+    points = np.random.rand(10,d)
     quadtree = tree.Points2Quadtree(points)
     qtree_out = quadtree.query_knn(x, k=k)
-    kdtree = scipy.spatial.KDTree(points)
-    kdtree_out = kdtree.query(x, k=k)
-    d = scipy.spatial.distance.cdist(points, np.atleast_2d(x)).flatten()
+    sci_kdtree = scipy.spatial.KDTree(points)
+    sci_kdtree_out = sci_kdtree.query(x, k=k)
+    dist = scipy.spatial.distance.cdist(points, np.atleast_2d(x)).flatten()
 
[…]
