### 2026-01-27 9f087d7
tests(tree): add tests for KNN queries
