### 2026-01-29 #9641 @jorgensd
> @nschloe, @jorgensd: Would either of you be interested? Danke and takk.

I can review this sometime in the end of next week
