### 2026-01-30 71f7498
fix(MarchingCubes): improve efficiency
