+++ src/mymesh/contour.py
@@ -4412,7 +4412,7 @@ def MarchingCubesImage(I, h=1, threshold=0, interpolation='linear', method='orig
             
     return NewCoords, NewConn
 
-def MarchingCubes(VoxelNodeCoords,VoxelNodeConn,NodeValues,threshold=0,interpolation='linear',method='33',flip=False, return_anchors=False):
+def MarchingCubes(VoxelNodeCoords,VoxelNodeConn,NodeValues,threshold=0,interpolation='linear',method='33',flip=False, return_anchors=False, cleanup_tol=1e-10, cleanup=True):
     """
     Marching cubes algorithm :cite:p:`Lorensen1987` for extracting an isosurface from a hexahedral mesh.
 
@@ -4440,6 +4440,12 @@ def MarchingCubes(VoxelNodeCoords,VoxelNodeConn,NodeValues,threshold=0,interpola
         algorithm or '33' for the marching cubes 33 algorithm :cite:p:`Chernyaev1995`.
     flip : bool, optional
         Flip the interior/exterior of the mesh, by default False. By default, values less than the threshold are assumed to be the “inside” of the mesh. If the inside is denoted by values greater than the threshold, set flip=True.
+    cleanup_tol : float, optional
+        Tolerance value used to classify whether two nodes are sufficiently close 
+        to be considered a single node (see :func:`mymesh.utils.DeleteDuplicateNodes`), 
+        by default 1e-10.
+    cleanup : bool, optional
+        If True, duplicate nodes will be merged, by default True.
 
     Returns
     -------
@@ -4467,7 +4473,7 @@ def MarchingCubes(VoxelNodeCoords,VoxelNodeConn,NodeValues,threshold=0,interpola
         _MarchingCubesLookup.LookupTable = MC_Lookup
     else:
         raise Exception('Unknown method "{:s}"'.format(method))
-    edgeLookup = [
+    edgeLookup = np.array([
         [0, 1],  # Edge 0 - Between nodes 0 and 1
         [1, 2],  # Edge 1
         [2, 3],  # Edge 2
@@ -4480,95 +4486,173 @@ def MarchingCubes(VoxelNodeCoords,VoxelNodeConn,NodeValues,threshold=0,interpola
         [5, 6],  # Edge 9
         [6, 7],  # Edge 10
         [7, 4],  # Edge 11
-        []       # Center
-        ]
-    arrayCoords = np.array(VoxelNodeCoords)
-    for e in range(len(VoxelNodeConn)):
-        vals = np.array([NodeValues[node] for node in VoxelNodeConn[e]])
-        inside = [1 if v <= 0 else 0 for v in vals]
-        i = int("".join(str(j) for j in inside), 2)
-        if method == 'original':
-            TriElems = _MarchingCubesLookup(i)
-        elif method == '33':
-            TriElems = _MarchingCubes33Lookup(i,vals)
-        else:
-            raise Exception('Invalid method')
+        [0, 6]   # Center
+        ])
+    # arrayCoords = np.asarray(VoxelNodeCoords)
+
+    HexVals = NodeValues[VoxelNodeConn]
+    inside = HexVals <= 0
+    ints = np.sum(inside * 2**np.arange(0,8)[::-1], axis=1)
+    if method.lower() == 'original' or method.lower() == 'mc':
+        flip = ints > 127
+        ints[flip] = 128 - (ints[flip] - 127)
+        element_lists = MC_Lookup[ints]
+    elif method.lower() == '33' or method.lower() == 'mc33':
+        # cases = MC33_Cases[ints]
+        # ambiguities = np.isin(cases, (3, 4, 6, 7, 10, 12, 13))
[…]
