### 2026-01-30 bb780c0
fix(SurfaceMesh): fix some options for MarchingCubes
