### 2026-02-06 e0d5bc5
fix(contour): additional efficiency improvements for contouring
