+++ src/mymesh/contour.py
@@ -27,6 +27,7 @@ Contouring
 import time
 import sys
 from . import utils, tree, converter
+from . import try_njit
 import numpy as np
 import scipy
 import copy, warnings
@@ -4250,11 +4251,12 @@ def MarchingCubesImage(I, h=1, threshold=0, interpolation='linear', method='orig
         tableIdx[flip] = 128 - (tableIdx[flip]-127)
         edgeList =  LookupTable[tableIdx]
         edgeConnections = np.array([x[::-1] if flip[i] else x for i,y in enumerate(edgeList) for x in y if len(x) != 0])
-        numbering = np.array([j for j,y in enumerate(edgeList) for i,x in enumerate(y) if len(x) != 0])
+        numbering = np.array([j for j,y in enumerate(edgeList) for i,x in enumerate(y) if x != []])
     
     elif method == '33':
         ##### TODO ####
-        pass
+        raise NotImplementedError('Marching Cubes 33 not yet implemented for MarchingCubesImage.')
+        
     if len(numbering) == 0:
         return np.empty((0,3)), np.empty((0,3))
     i_indices = icubes[numbering][np.arange(len(numbering))[:, np.newaxis, np.newaxis], edgeLookup[edgeConnections]]
@@ -4412,7 +4414,7 @@ def MarchingCubesImage(I, h=1, threshold=0, interpolation='linear', method='orig
             
     return NewCoords, NewConn
 
-def MarchingCubes(VoxelNodeCoords,VoxelNodeConn,NodeValues,threshold=0,interpolation='linear',method='33',flip=False, return_anchors=False, cleanup_tol=1e-10, cleanup=True):
+def MarchingCubes(VoxelNodeCoords,VoxelNodeConn,NodeValues,threshold=0,interpolation='linear',method='33',flip=False, return_anchors=False, cleanup_tol=1e-10, cleanup=False):
     """
     Marching cubes algorithm :cite:p:`Lorensen1987` for extracting an isosurface from a hexahedral mesh.
 
@@ -4461,18 +4463,18 @@ def MarchingCubes(VoxelNodeCoords,VoxelNodeConn,NodeValues,threshold=0,interpola
     Anchors = []
     AnchorAxis = []
     AnchorDir = []
-    NodeValues = np.asarray(NodeValues) - threshold
+    NodeValues = np.asarray(NodeValues,dtype=float) - threshold
     if flip:
         NodeValues = -1*NodeValues
-    if method == '33':
-        _MarchingCubes33Lookup.LookupTable = MC33_Lookup
-        _MarchingCubes33Lookup.FaceTests = MC33_FaceTest
-        _MarchingCubes33Lookup.Cases = MC33_Cases
-        _MarchingCubes33Lookup.Signs = MC33_Signs
-    elif method == 'original':
-        _MarchingCubesLookup.LookupTable = MC_Lookup
-    else:
-        raise Exception('Unknown method "{:s}"'.format(method))
+    # if method == '33':
+    #     _MarchingCubes33Lookup.LookupTable = MC33_Lookup
+    #     _MarchingCubes33Lookup.FaceTests = MC33_FaceTest
+    #     _MarchingCubes33Lookup.Cases = MC33_Cases
+    #     _MarchingCubes33Lookup.Signs = MC33_Signs
+    # elif method == 'original':
+    #     _MarchingCubesLookup.LookupTable = MC_Lookup
+    # else:
+    #     raise Exception('Unknown method "{:s}"'.format(method))
     edgeLookup = np.array([
         [0, 1],  # Edge 0 - Between nodes 0 and 1
         [1, 2],  # Edge 1
@@ -4498,27 +4500,31 @@ def MarchingCubes(VoxelNodeCoords,VoxelNodeConn,NodeValues,threshold=0,interpola
[…]

[…]
