### 2026-02-11 #9641 @jorgensd
I've started looking through the paper and code today, and will hopefully have the first set of review comments ready tomorrow.
