### 2026-02-15 #10003 @editorialbot
[REVIEW]: MyMesh: General purpose, implicit, and image-based meshing in python
**Submitting author:** @TimJosephson (<a href="http://orcid.org/0000-0003-2703-6487">Timothy Josephson</a>)
**Repository:** https://github.com/BU-SMBL/mymesh
**Branch with paper.md** (empty if default branch): paper
**Version:** 0.3.0
