### 2026-02-17 2f19b5e
feat(contour): add voxel_mode option to SurfaceMesh and TetMesh
