### 2026-02-17 4ca17a4
fix(image): reduce number of unnecessary voxels going to marching cubes/tets
