### 2026-02-23 #10003 @TimJosephson
@RichardLitt, in order to address some comments from the reviewers, it will be necessary to make some changes that would amount to a new version (likely 0.4.0 instead of 0.3.0). Is there any preference for how such changes are implemented during the review process, since the review is currently associated with version 0.3.0 (e.g. I could make a new release, or I could keep the changes in the `paper` branch until the end of the review)? The changes would likely also include some fixes/features that have been developed since the time of submission that are not directly tied to reviewer comments, but are nevertheless relevant to some comments related to performance/benchmarking and functionality. However, I also want to be mindful of not complicating things by making changes to the content being reviewed in the middle of the review. Please let me know if you (or the reviewers) have any preferences for how such changes are handled. Thanks!
