### 2026-02-25 #10003 @RichardLitt
My preference is to prioritize the health of your repository over the editorial needs of the paper. It's fine to make changes in response to the review during the review; this is a dynamic process. I would suggest you link to this thread in any updates which are relevant to your repo; that way, reviewers can see how you're responding to their comments. We can repin the version association with the paper when needed.
