### 2026-02-25 ba5659b
fix(contour): bug fix in MarchingElements
