### 2026-02-25 7b445d8
fix(utils.Face2NodeNormal): fix nan handling
