### 2026-02-25 ce4bacf
docs: set up version switcher
