### 2026-02-25 fa0c813
docs: update version switcher
