+++ 0.1.1/_sources/generated/submodules/mymesh.register.rotation.rst.txt
@@ -1,0 +1,6 @@
+mymesh.register.rotation
+========================
+
+.. currentmodule:: mymesh.register
+
+.. autofunction:: rotation
