+++ 0.1.1/_sources/generated/submodules/mymesh.tree.OctreeNode.ContainsPts.rst.txt
@@ -1,0 +1,6 @@
+mymesh.tree.OctreeNode.ContainsPts
+==================================
+
+.. currentmodule:: mymesh.tree
+
+.. automethod:: OctreeNode.ContainsPts
