+++ 0.1.1/_sources/generated/submodules/mymesh.improvement.TetFlip.rst.txt
@@ -1,0 +1,6 @@
+mymesh.improvement.TetFlip
+==========================
+
+.. currentmodule:: mymesh.improvement
+
+.. autofunction:: TetFlip
