+++ 0.1.1/_sources/generated/submodules/mymesh.utils.MergeMesh.rst.txt
@@ -1,0 +1,6 @@
+mymesh.utils.MergeMesh
+======================
+
+.. currentmodule:: mymesh.utils
+
+.. autofunction:: MergeMesh
