+++ 0.1.1/_sources/generated/submodules/mymesh.converter.wdg152linear.rst.txt
@@ -1,0 +1,6 @@
+mymesh.converter.wdg152linear
+=============================
+
+.. currentmodule:: mymesh.converter
+
+.. autofunction:: wdg152linear
