+++ 0.1.1/_sources/generated/submodules/mymesh.quality.Skewness.rst.txt
@@ -1,0 +1,6 @@
+mymesh.quality.Skewness
+=======================
+
+.. currentmodule:: mymesh.quality
+
+.. autofunction:: Skewness
