+++ 0.1.1/_sources/generated/submodules/mymesh.converter.solid2surface.rst.txt
@@ -1,0 +1,6 @@
+mymesh.converter.solid2surface
+==============================
+
+.. currentmodule:: mymesh.converter
+
+.. autofunction:: solid2surface
