+++ 0.1.1/_sources/generated/mymesh.mesh.mesh.NFace.rst.txt
@@ -1,0 +1,6 @@
+mymesh.mesh.mesh.NFace
+======================
+
+.. currentmodule:: mymesh.mesh
+
+.. autoproperty:: mesh.NFace
