### 2026-02-25 37d1571
docs: Add old docs version 0.2.0
