+++ 0.2.0/_sources/generated/submodules/mymesh.implicit.neovius.rst.txt
@@ -1,0 +1,6 @@
+mymesh.implicit.neovius
+=======================
+
+.. currentmodule:: mymesh.implicit
+
+.. autofunction:: neovius
