+++ 0.2.0/_sources/generated/submodules/mymesh.implicit.grid2fun.rst.txt
@@ -1,0 +1,6 @@
+mymesh.implicit.grid2fun
+========================
+
+.. currentmodule:: mymesh.implicit
+
+.. autofunction:: grid2fun
