+++ 0.2.0/_sources/generated/mymesh.mesh.mesh.Contour.rst.txt
@@ -1,0 +1,6 @@
+mymesh.mesh.mesh.Contour
+========================
+
+.. currentmodule:: mymesh.mesh
+
+.. automethod:: mesh.Contour
