+++ 0.2.0/_sources/generated/submodules/mymesh.register.ImageOverlay.rst.txt
@@ -1,0 +1,6 @@
+mymesh.register.ImageOverlay
+============================
+
+.. currentmodule:: mymesh.register
+
+.. autofunction:: ImageOverlay
