+++ 0.2.0/_sources/generated/submodules/mymesh.implicit.gyroid.rst.txt
@@ -1,0 +1,6 @@
+mymesh.implicit.gyroid
+======================
+
+.. currentmodule:: mymesh.implicit
+
+.. autofunction:: gyroid
