+++ 0.2.0/_sources/generated/submodules/mymesh.register.similarity2d.rst.txt
@@ -1,0 +1,6 @@
+mymesh.register.similarity2d
+============================
+
+.. currentmodule:: mymesh.register
+
+.. autofunction:: similarity2d

[…]
