### 2026-02-25 e414c9d
docs: Add old docs version 0.3.0
