+++ 0.3.0/_sources/generated/submodules/mymesh.improvement.TaubinSmoothing.rst.txt
@@ -1,0 +1,6 @@
+﻿mymesh.improvement.TaubinSmoothing
+==================================
+
+.. currentmodule:: mymesh.improvement
+
+.. autofunction:: TaubinSmoothing
