+++ 0.3.0/_sources/generated/submodules/mymesh.utils.CleanupDegenerateElements.rst.txt
@@ -1,0 +1,6 @@
+mymesh.utils.CleanupDegenerateElements
+======================================
+
+.. currentmodule:: mymesh.utils
+
+.. autofunction:: CleanupDegenerateElements
