+++ 0.3.0/_sources/generated/submodules/mymesh.quality.tet_circumradius.rst.txt
@@ -1,0 +1,6 @@
+mymesh.quality.tet\_circumradius
+================================
+
+.. currentmodule:: mymesh.quality
+
+.. autofunction:: tet_circumradius
