+++ 0.3.0/_sources/generated/submodules/mymesh.utils.Centroids.rst.txt
@@ -1,0 +1,6 @@
+mymesh.utils.Centroids
+======================
+
+.. currentmodule:: mymesh.utils
+
+.. autofunction:: Centroids
