+++ 0.3.0/_sources/generated/submodules/mymesh.improvement.Split.rst.txt
@@ -1,0 +1,6 @@
+﻿mymesh.improvement.Split
+========================
+
+.. currentmodule:: mymesh.improvement
+
+.. autofunction:: Split
