+++ 0.3.0/_sources/generated/submodules/mymesh.rays.PointsInVoxel.rst.txt
@@ -1,0 +1,6 @@
+mymesh.rays.PointsInVoxel
+=========================
+
+.. currentmodule:: mymesh.rays
+
+.. autofunction:: PointsInVoxel
