+++ 0.3.0/_sources/generated/submodules/mymesh.rays.PointInTri.rst.txt
@@ -1,0 +1,6 @@
+mymesh.rays.PointInTri
+======================
+
+.. currentmodule:: mymesh.rays
+
+.. autofunction:: PointInTri
