+++ 0.3.0/_sources/generated/mymesh.mesh.mesh.MeshNodes.rst.txt
@@ -1,0 +1,6 @@
+mymesh.mesh.mesh.MeshNodes
+==========================
+
+.. currentmodule:: mymesh.mesh
+
+.. autoproperty:: mesh.MeshNodes
