+++ 0.3.0/_sources/examples/index.rst.txt
@@ -1,0 +1,140 @@
+:orphan:
+
+Examples
+========
+
+Examples highlighting the capabilities of MyMesh
+
+
+.. raw:: html
+
+    <div class="sphx-glr-thumbnails">
+
+.. thumbnail-parent-div-open
+
+.. thumbnail-parent-div-close
+
+.. raw:: html
+
+    </div>
+
+Mesh Analysis Examples
+======================
+
+Examples for analyzing meshes.
+
+
+.. raw:: html
+
+    <div class="sphx-glr-thumbnails">
+
+.. thumbnail-parent-div-open
+
+.. raw:: html
+
+    <div class="sphx-glr-thumbcontainer" tooltip="Curvature Analysis">
+
+.. only:: html
+
+  .. image:: /examples/MeshAnalysis/images/thumb/sphx_glr_demo_curvature_thumb.png
+    :alt:
+
+  :ref:`sphx_glr_examples_MeshAnalysis_demo_curvature.py`
+
+.. raw:: html
+
+      <div class="sphx-glr-thumbnail-title">Curvature Analysis</div>
+    </div>
+
+
+.. thumbnail-parent-div-close
+
+.. raw:: html
+
+    </div>
+
+Mesh Generation Examples
+========================
+
+Examples for different methods of generating meshes.
+
+
+.. raw:: html
+
+    <div class="sphx-glr-thumbnails">
+
+.. thumbnail-parent-div-open
+
+.. raw:: html
+
+    <div class="sphx-glr-thumbcontainer" tooltip="This example generates a surface mesh of an implicit function (Taubin&#x27;s heart surface :citeTaubin1993) and then performs  Laplacian smoothing on it.">
+
+.. only:: html
+
+  .. image:: /examples/MeshGeneration/images/thumb/sphx_glr_demo_heart_thumb.png
+    :alt:
+
+  :ref:`sphx_glr_examples_MeshGeneration_demo_heart.py`
+
+.. raw:: html
+
+      <div class="sphx-glr-thumbnail-title">Implicit Heart</div>
+    </div>
+
+
+.. raw:: html
+
+    <div class="sphx-glr-thumbcontainer" tooltip="This example creates pasta shapes with mymesh.primitives using sweep  construction methods (~mymesh.primitives.Extrude and  Revolve). These demonstrate how meshes  of complex shapes can be constructed starting with a single straight line.">
+
+.. only:: html
+
+  .. image:: /examples/MeshGeneration/images/thumb/sphx_glr_demo_pasta_thumb.png
+    :alt:
+
+  :ref:`sphx_glr_examples_MeshGeneration_demo_pasta.py`
+
+.. raw:: html
+
+      <div class="sphx-glr-thumbnail-title">Pasta Shapes</div>
+    </div>
+
+
+.. raw:: html
+
+    <div class="sphx-glr-thumbcontainer" tooltip="Triply periodic minimal surfaces (TPMS) are particularly interesting subjects for a mixed-topology approach because the resultant surface will remain triply periodic (though not minimal) as long as the periodicities are compatible.  ">
+
+.. only:: html
+
+  .. image:: /examples/MeshGeneration/images/thumb/sphx_glr_demo_mixed-top_thumb.png
+    :alt:
+
+  :ref:`sphx_glr_examples_MeshGeneration_demo_mixed-top.py`
+
+.. raw:: html
+
+      <div class="sphx-glr-thumbnail-title">Mixed-Topology TPMS</div>
+    </div>
+
+
+.. thumbnail-parent-div-close
+
+.. raw:: html
+
+    </div>
+
+
+.. toctree::
+   :hidden:
+   :includehidden:
+
+
[…]
