+++ 0.3.0/_sources/generated/submodules/mymesh.implicit.unionf.rst.txt
@@ -1,0 +1,6 @@
+mymesh.implicit.unionf
+======================
+
+.. currentmodule:: mymesh.implicit
+
+.. autofunction:: unionf
