### 2026-02-25 903e848
docs: update version switcher
