### 2026-02-26 bd256b1
docs: Update old docs version 0.1.1
