+++ 0.1.1/_sources/examples/MeshGeneration/demo_heart.rst.txt
@@ -123,7 +123,7 @@ the mesh and more evenly distribute the nodes.
 
 .. rst-class:: sphx-glr-timing
 
-   **Total running time of the script:** (0 minutes 0.696 seconds)
+   **Total running time of the script:** (0 minutes 0.706 seconds)
 
 
 .. _sphx_glr_download_examples_MeshGeneration_demo_heart.py:
