+++ 0.1.1/_sources/examples/MeshAnalysis/sg_execution_times.rst.txt
@@ -6,7 +6,7 @@
 
 Computation times
 =================
-**00:15.923** total execution time for 1 file **from examples/MeshAnalysis**:
+**00:16.656** total execution time for 1 file **from examples/MeshAnalysis**:
 
 .. container::
 
@@ -33,5 +33,5 @@ Computation times
      - Time
      - Mem (MB)
    * - :ref:`sphx_glr_examples_MeshAnalysis_demo_curvature.py` (``demo_curvature.py``)
-     - 00:15.923
+     - 00:16.656
      - 0.0
