### 2026-02-26 c6c7e3c
docs: Update old docs version 0.2.0
