+++ 0.2.0/_sources/examples/MeshAnalysis/demo_curvature.rst.txt
@@ -184,7 +184,7 @@ curvatures deviate from the true curvature of 1 mm :sup:`-1`.
 
 .. rst-class:: sphx-glr-timing
 
-   **Total running time of the script:** (0 minutes 15.923 seconds)
+   **Total running time of the script:** (0 minutes 16.653 seconds)
 
 
 .. _sphx_glr_download_examples_MeshAnalysis_demo_curvature.py:
