+++ 0.2.0/_sources/examples/MeshGeneration/demo_pasta.rst.txt
@@ -125,7 +125,7 @@ Please submit any new pasta constructions to toj@bu.edu
 
 .. rst-class:: sphx-glr-timing
 
-   **Total running time of the script:** (0 minutes 4.677 seconds)
+   **Total running time of the script:** (0 minutes 4.906 seconds)
 
 
 .. _sphx_glr_download_examples_MeshGeneration_demo_pasta.py:
