+++ 0.2.0/_sources/examples/MeshGeneration/demo_mixed-top.rst.txt
@@ -129,7 +129,7 @@ including :func:`~mymesh.implicit.gyroid`, :func:`~mymesh.implicit.primitive`
 
 .. rst-class:: sphx-glr-timing
 
-   **Total running time of the script:** (0 minutes 12.941 seconds)
+   **Total running time of the script:** (0 minutes 8.569 seconds)
 
 
 .. _sphx_glr_download_examples_MeshGeneration_demo_mixed-top.py:
