### 2026-02-26 d2c8d2b
docs: Update old docs version 0.3.0
