+++ 0.3.0/_sources/generated/mymesh.mesh.mesh.Transform.rst.txt
@@ -1,4 +1,4 @@
-﻿mymesh.mesh.mesh.Transform
+mymesh.mesh.mesh.Transform
 ==========================
 
 .. currentmodule:: mymesh.mesh
