+++ 0.3.0/_sources/generated/submodules/mymesh.improvement.TaubinSmoothing.rst.txt
@@ -1,4 +1,4 @@
-﻿mymesh.improvement.TaubinSmoothing
+mymesh.improvement.TaubinSmoothing
 ==================================
 
 .. currentmodule:: mymesh.improvement
