+++ 0.3.0/_sources/generated/submodules/mymesh.improvement.Split.rst.txt
@@ -1,4 +1,4 @@
-﻿mymesh.improvement.Split
+mymesh.improvement.Split
 ========================
 
 .. currentmodule:: mymesh.improvement
