+++ 0.3.0/_sources/examples/MeshAnalysis/sg_execution_times.rst.txt
@@ -6,7 +6,7 @@
 
 Computation times
 =================
-**03:27.396** total execution time for 1 file **from examples/MeshAnalysis**:
+**04:05.554** total execution time for 1 file **from examples/MeshAnalysis**:
 
 .. container::
 
@@ -33,5 +33,5 @@ Computation times
      - Time
      - Mem (MB)
    * - :ref:`sphx_glr_examples_MeshAnalysis_demo_curvature.py` (``demo_curvature.py``)
-     - 03:27.396
+     - 04:05.554
      - 0.0
