+++ 0.3.0/_sources/examples/MeshGeneration/demo_pasta.rst.txt
@@ -115,17 +115,25 @@ Please submit any new pasta constructions to toj@bu.edu
  .. code-block:: none
 
     RFBOutputContext()
-    RFBOutputContext()
-    RFBOutputContext()
-    RFBOutputContext()
 
+    Identifying mesh nodes...Done
+            RFBOutputContext()
+
+    Identifying mesh nodes...Done
+            RFBOutputContext()
+
+    Identifying mesh nodes...Done
+            RFBOutputContext()
+
+    Identifying mesh nodes...Done
+        
 
 
 
 
 .. rst-class:: sphx-glr-timing
 
-   **Total running time of the script:** (0 minutes 4.677 seconds)
+   **Total running time of the script:** (0 minutes 5.328 seconds)
 
 
 .. _sphx_glr_download_examples_MeshGeneration_demo_pasta.py:
