+++ 0.3.0/_sources/examples/MeshAnalysis/demo_curvature.rst.txt
@@ -273,7 +273,7 @@ and then evaluated at the nodes of the mesh.
 
 .. rst-class:: sphx-glr-timing
 
-   **Total running time of the script:** (3 minutes 27.396 seconds)
+   **Total running time of the script:** (4 minutes 5.554 seconds)
 
 
 .. _sphx_glr_download_examples_MeshAnalysis_demo_curvature.py:
