### 2026-02-26 6737c99
docs: documentation fixes and touch ups
