+++ src/mymesh/converter.py
@@ -317,27 +317,27 @@ def solid2edges(NodeCoords,NodeConn,ElemType='auto',return_EdgeConn=False,return
     ElemType : str, optional
         Specifies the element type contained within the mesh, by default 'auto'.
 
-        - 'auto' or 'mixed' - Will detect element type by the number of nodes present in each element using :func:`~mymesh.utils.identify_type`. 
+        - 'auto' or 'mixed': Will detect element type by the number of nodes present in each element using :func:`~mymesh.utils.identify_type`. 
 
-        - 'surf' - Will detect element type by the number of nodes present in each 
+        - 'surf': Will detect element type by the number of nodes present in each 
         element, assuming four node elements are quads
 
-        - 'vol' - Will detect element type by the number of nodes present in each 
+        - 'vol': Will detect element type by the number of nodes present in each 
         element, assuming four node elements are tets (functionally the ame as 'auto')
 
-        - 'tri' - All elements treated as 3-node triangular elements.
+        - 'tri': All elements treated as 3-node triangular elements.
 
-        - 'quad' - All elements treated as 4-node quadrilateral elements.
+        - 'quad': All elements treated as 4-node quadrilateral elements.
 
-        - 'tet' - All elements treated as 4-node tetrahedral elements.
+        - 'tet': All elements treated as 4-node tetrahedral elements.
 
-        - 'pyramid' - All elements treated as 5-node wedge elements.
+        - 'pyramid': All elements treated as 5-node wedge elements.
 
-        - 'wedge' - All elements treated as 6-node quadrilateral elements.
+        - 'wedge': All elements treated as 6-node quadrilateral elements.
 
-        - 'hex' - All elements treated as 8-node quadrilateral elements.
+        - 'hex': All elements treated as 8-node quadrilateral elements.
 
-        - 'polygon' - All elements treated as n-node polygonal elements. TODO: add support for return_EdgeConn and return_EdgeElem
+        - 'polygon': All elements treated as n-node polygonal elements. TODO: add support for return_EdgeConn and return_EdgeElem
 
     return_EdgeConn : bool, optional
         If true, will return EdgeConn, the Edge Connectivity of each element.
@@ -534,7 +534,7 @@ def EdgesByElement(NodeCoords,NodeConn,ElemType='auto'):
     ElementEdges : list
         Edge connectivity, grouped by element
     """    
-    Edges, EdgeConn = converter.solid2edges(*S, return_EdgeConn=True, ElemType=ElemType)
+    Edges, EdgeConn = solid2edges(NodeCoords, NodeConn, return_EdgeConn=True, ElemType=ElemType)
     ElementEdges = [Edges[ec] for ec in EdgeConn]
     return ElementEdges
 
