### 2026-02-26 95454a7
fix(utils.BaryTri): Improve barycentric coordinate efficiency
