+++ src/mymesh/utils.py
@@ -664,113 +664,10 @@ def Face2NodeNormal(NodeCoords,NodeConn,ElemConn,ElemNormals,method='Angle'):
 
     else:
         raise ValueError(f'Invalid method: {method:s}')
-        # NodeNormals = [[] for i in range(len(NodeCoords))]      # Normal vectors for each vertex
-        # NodeSet = {n for elem in NodeConn for n in elem}
-        # for i in range(len(NodeCoords)):
-        #     if i not in NodeSet:
-        #         NodeNormals[i] = [np.nan,np.nan,np.nan]
-        #         continue
-        #     angles = [0 for j in range(len(ElemConn[i]))]
-        #     elemnormals = [np.array(ElemNormals[elem]) for elem in ElemConn[i]]
-
-        #     if method == 'MostVisible_Loop':
-                
-        #         # This is kept for readability; 'MostVisible' is a vectorized equivalent that performs significantly faster
-                
-        #         # Note: this code uses dot(Ni,Nj) as a surrogate for radius; since Ni,Nj are both unit vectors
-        #         # cos(theta) = dot(Ni,Nj) -> theta = arccos(dot(Ni,Nj)). Since arccos is a monotonically 
-        #         # decreasing function, if dot(Ni,Nj) < dot(Ni,Nk), then rij > rik
-        #         eps = -1e-8
-        #         scalmin = -1
-        #         C = [np.nan,np.nan,np.nan]
-        #         for ii in range(len(elemnormals)-1):
-        #             # Check the 2 point circles
-        #             Ni = np.array(elemnormals[ii])
-        #             for j in range(ii+1,len(elemnormals)):
-        #                 Nj = np.array(elemnormals[j])
-        #                 Nb = Ni+Nj
-        #                 Nb = Nb/np.linalg.norm(Nb)
-        #                 scal = np.dot(Nb,Ni)
-        #                 if scal < scalmin:      
-        #                     pass
-        #                 elif any((np.dot(Nl,Nb) - scal) < eps for Nl in elemnormals):
-        #                     pass
-        #                 else:
-        #                     C = Nb.tolist()
-        #                     scalmin = scal
-        #         for ii in range(len(elemnormals)-2): 
-        #             # Check the 3 point circles
-        #             Ni = elemnormals[ii]
-        #             for j in range(ii+1,len(elemnormals)-1):
-        #                 Nj = elemnormals[j]
-        #                 for k in range(j+1,len(elemnormals)):
-        #                     Nk = elemnormals[k]
-
-        #                     denom = (2*np.linalg.norm(np.cross(Ni-Nk,Nj-Nk))**2) 
-        #                     if denom == 0:
-        #                         continue
-        #                     Nc = np.cross(np.linalg.norm(Ni-Nk)**2 * (Nj-Nk) - np.linalg.norm(Nj-Nk)**2 * (Ni-Nk), np.cross(Ni-Nk,Nj-Nk))/denom + Nk
-        #                     nNc = np.linalg.norm(Nc)
-        #                     if nNc == 0:
-        #                         continue
-        #                     Nc = Nc/nNc
-                            
-
-        #                     scal = np.dot(Nc, Ni)
-        #                     if scal < 0:
-        #                         Nc = [-1*n for n in Nc]
-        #                         scal = -scal
[…]
