### 2026-02-26 14a9360
fix(mymesh.dmesh): convert from jitclass to StructRef
