+++ src/mymesh/mesh.py
@@ -9,7 +9,7 @@ mesh class
 
 """
 
-from . import utils, implicit, improvement, contour, converter, quality, rays, register, curvature, visualize, check_numba
+from . import utils, implicit, improvement, contour, converter, quality, rays, register, curvature, visualize, check_numba, try_njit
 from sys import getsizeof
 import scipy
 import numpy as np
@@ -2562,9 +2562,7 @@ class mesh:
         D : mesh.dmesh
             dynamic mesh
 
-
         """      
-
         # Ensure relevant data is in properly typed numpy arrays
         try:
             NodeConn = np.array(self.NodeConn, dtype=np.int64)
@@ -2573,66 +2571,303 @@ class mesh:
         
         NodeCoords = np.array(self.NodeCoords, dtype=np.float64)
 
-        Edges = np.array(self.Edges, dtype=np.int64)
-
-        try:
-            Faces = np.array(self.Faces, dtype=np.int64)
-        except:
-            raise ValueError('Dynamic meshes are not valid for meshes with elements that have mixed faces (e.g. wedges or pyramids).')
-        
-        # ElemNeighbors = utils.PadRagged(self.ElemNeighbors. fillval=-1)
-        
-        ElemConn_head = np.full(self.NNode, -1, dtype=np.int64)
-        ElemConn_elem = np.array([e for elems in self.ElemConn for e in elems], dtype=np.int64)
-        ElemConn_next = np.full(len(ElemConn_elem), -1, dtype=np.int64)
-        ElemConn_prev = np.full(len(ElemConn_elem), -1, dtype=np.int64)
-        ElemConn_tail = np.full(self.NNode, -1, dtype=np.int64)
-        k = 0
-        for i,elems in enumerate(self.ElemConn):
-            if len(elems) == 0:
-                continue
-            ElemConn_head[i] = k
-            # ElemConn_prev[k] = -1 # Implied by initialization
-            for e in elems[:-1]:
-                ElemConn_next[k] = k+1
-                ElemConn_prev[k+1] = k
-                k += 1
-            # ElemConn_next[k] = -1 # Implied by initialization
-            ElemConn_tail[i] = k
-            k += 1
-        
-        # if self.Type == 'surf':
-        #     # Neighbors are connected by edges
-        #     NeighborConnection = len(Edges[0])
-        # elif self.Type == 'vol':
-        #     # Neighbors are connected by faces
-        #     NeighborConnection = len(Faces[0])
         if ElemLabels is None:
             ElemLabels = np.empty(0, np.int64)
         else:
             ElemLabels = np.asarray(ElemLabels, np.int64)
+        if self.verbose:
+            self.verbose = False
+            _ = self.ElemConn
+            self.verbose = True
 
-        D = dmesh(NodeCoords, NodeConn, ElemConn_head, ElemConn_elem, ElemConn_next, ElemConn_prev, ElemConn_tail, ElemLabels)
+        D = dmesh(NodeCoords, NodeConn, self.ElemConn, ElemLabels)
 
         return D
+    
+# dmesh class
+# Attributes:
+@try_njit(cache=True)
+def dmesh_get_raw_NodeCoords(self):
+    return self.raw_NodeCoords
+@try_njit(cache=True)
+def dmesh_get_NNode(self):
+    return self.NNode  
+@try_njit(cache=True)
+def dmesh_get_NodeCoords(self):
+    return self.raw_NodeCoords[:self.NNode]
+@try_njit(cache=True)
+def dmesh_get_raw_NodeConn(self):
+    return self.raw_NodeConn
+@try_njit(cache=True)
+def dmesh_get_NElem(self):
+    return self.NElem
[…]
