### 2026-02-26 8395fce
fix(delaunay.BowyerWatson2d): major efficiency improvements
