+++ src/mymesh/delaunay.py
@@ -293,10 +293,11 @@ def Triangle(NodeCoords,Constraints=None, steiner=0):
     # Uses Triangle by Jonathan Shewchuk
     if Constraints is None or len(Constraints)==0:
         In = dict(vertices=NodeCoords)
+        Out = triangle.triangulate(In,f'qS{steiner:d}')
     else:
         In = dict(vertices=NodeCoords,segments=Constraints)
-    try:
         Out = triangle.triangulate(In,f'pqS{steiner:d}')
+    try:
         NodeConn = Out['triangles']
         # NodeCoords = Out['vertices']
         if len(Out['vertices']) != len(NodeCoords):
@@ -309,6 +310,7 @@ def Triangle(NodeCoords,Constraints=None, steiner=0):
             if np.any(NodeConn >= len(NodeCoords)):
                 NodeCoords = Out['vertices']
     except:
+        warnings.warn('Error using Triangle, falling back to SciPy')
         NodeConn = SciPy(NodeCoords)
 
     return NodeCoords, NodeConn
@@ -491,7 +493,7 @@ def TriangleSplitting(NodeCoords, Hull=None):
 
     return NodeConn
         
-def BowyerWatson2d(NodeCoords):
+def BowyerWatson2d(NodeCoords, nsample=3):
     """
     Bowyer-Watson algorithm for 2D Delaunay triangulation
 
@@ -508,12 +510,8 @@ def BowyerWatson2d(NodeCoords):
     NodeConn : np.ndarray
         mx3 array of node connectivities for the Delaunay triangulation
     """
-    if check_numba():
-        import numba
-        from numba.typed import Dict
-    else:
-        warnings.warn('Using numba is strongly recommended for efficiency of BowyerWatson2d. Activate with `mymesh.use_numba(True)`')
-        Dict = dict
+    if not check_numba():
+        warnings.warn('Using numba is strongly recommended for efficiency of BowyerWatson2d.')
 
     NodeCoords = np.asarray(NodeCoords)
     assert NodeCoords.shape[0] >= 3, 'At least three points are required.'
@@ -526,14 +524,15 @@ def BowyerWatson2d(NodeCoords):
     nPts = len(NodeCoords)
 
     # Random insertion order for points
-    indices = list(range(nPts))
-    rng = np.random.default_rng()
-    rng.shuffle(indices)
+    # indices = list(range(nPts))
+    # rng = np.random.default_rng()
+    # rng.shuffle(indices)
+    indices = _bin_sort_2d(NodeCoords)
 
     # Get super triangle - triangle with incircle that bounds the point set
     center = np.mean(TempCoords, axis=0)
     r = np.max(np.sqrt((TempCoords[:,0]-center[0])**2 + (TempCoords[:,1]-center[1])**2))
-    R = r + 1*r/10
+    R = 10*r
 
     super_triangle_points = np.array([
                                     [center[0], center[1]-2*R],
@@ -543,41 +542,20 @@ def BowyerWatson2d(NodeCoords):
     TempCoords = np.hstack([np.vstack([TempCoords, super_triangle_points]), np.repeat(0,nPts+3)[:,None]])
     super_tri = (nPts, nPts+1, nPts+2)
 
-    ElemTable = Dict()
-    # Elem table links elements to tuples of (oriented) edges
-    # e.g. ElemTable[(0,1,2)] = ((0,1),(1,2),(2,0))
-    ElemTable[super_tri] = ((nPts, nPts+1), (nPts+1, nPts+2), (nPts+2, nPts)) 
+    
+    if 'mesh' in dir(mesh):
+        m = mesh.mesh(TempCoords, [super_tri], Type='surf', verbose=False)
+    else:
+        m = mesh(TempCoords, [super_tri], Type='surf', verbose=False)
+    
[…]
