### 2026-02-26 a87e550
fix(improvement): update Contract, Split, Flip to new dmesh class, other bug fixes
