+++ src/mymesh/improvement.py
@@ -31,8 +31,7 @@ Local mesh topology
 
     Contract
     Split
-    TetFlip
-    TetImprove
+    Flip
 
 
 """
@@ -1644,29 +1643,54 @@ def Contract(M, h, FixedNodes=set(), verbose=True, cleanup=True, labels=None, Fe
                 raise ValueError('If provided as a string, labels must correspond to an entry in M.ElemData')
         else:
             label_str = 'labels'
+            M.ElemData[label_str] = labels
         assert len(labels) == M.NElem, 'labels must correspond to the number of elements.'
-        if 'mesh' in dir(mesh):
-            MultiSurface = mesh.mesh(verbose=False)
-        else:
-            MultiSurface = mesh(verbose=False)
-        ulabels = np.unique(labels)
-        label_nodes = np.zeros((len(ulabels),M.NNode),dtype=int) # For each label, boolean indicator of whether each node is touching an element with that label
-        mesh_nodes = np.arange(M.NNode)
-        for i,label in enumerate(ulabels):
+        if M.Type == 'vol':
             if 'mesh' in dir(mesh):
-                m = mesh.mesh(M.NodeCoords, M.NodeConn[labels == label], verbose=False)
+                MultiSurface = mesh.mesh(verbose=False)
             else:
-                m = mesh(M.NodeCoords, M.NodeConn[labels == label], verbose=False)
-            
-            MultiSurface.addElems(m.SurfConn)
-            label_nodes[i,np.unique(m.NodeConn)] = 1
-        MultiSurface.NodeCoords = M.NodeCoords    
-        MultiSurface.Type = 'surf'
-        MultiSurface.NodeConn = MultiSurface.Faces  # This prevents doubling of surface elements at interfaces
-        SurfConn = MultiSurface.NodeConn
-        SurfEdges = np.sort(MultiSurface.Edges)
-        JunctionEdges = MultiSurface.Edges[np.array([len(conn) > 2 for conn in MultiSurface.EdgeElemConn])]
-        JunctionNodes = np.unique(JunctionEdges)
+                MultiSurface = mesh(verbose=False)
+            ulabels = np.unique(labels)
+            label_nodes = np.zeros((len(ulabels),M.NNode),dtype=int) # For each label, boolean indicator of whether each node is touching an element with that label
+            mesh_nodes = np.arange(M.NNode)
+            for i,label in enumerate(ulabels):
+                if 'mesh' in dir(mesh):
+                    m = mesh.mesh(M.NodeCoords, np.asarray(M.NodeConn)[labels == label], verbose=False)
+                else:
+                    m = mesh(M.NodeCoords, np.asarray(M.NodeConn)[labels == label], verbose=False)
+                
+                MultiSurface.addElems(m.SurfConn)
+                label_nodes[i,np.unique(m.NodeConn)] = 1
+            MultiSurface.NodeCoords = M.NodeCoords    
+            MultiSurface.Type = 'surf'
+            MultiSurface.NodeConn = MultiSurface.Faces  # This prevents doubling of surface elements at interfaces
+            SurfConn = MultiSurface.NodeConn
+            SurfEdges = np.sort(MultiSurface.Edges)
+            JunctionEdges = MultiSurface.Edges[np.array([len(conn) > 2 for conn in MultiSurface.EdgeElemConn])]
+            JunctionNodes = np.unique(JunctionEdges)
+        elif M.Type == 'surf':
+            if 'mesh' in dir(mesh):
[…]
