### 2026-02-26 4d27c5a
fix(__init__): Update optional numba logic, export dmesh
