### 2026-02-28 f3aaa6a
fix(contour.MarchingCubesImage): fix bug in cleanup
