### 2026-02-28 e4cde8b
fix(primitives.Revolve): fix occasional overlapping elements
