### 2026-02-28 54bdd38
feat(mesh.plot): add optional input for title
