+++ tests/test_improvement.py
@@ -1,6 +1,6 @@
 import pytest
 import numpy as np
-from mymesh import improvement, mesh
+from mymesh import improvement, mesh, implicit, quality, primitives
 
 @pytest.mark.parametrize("M, options", [
     # Case 1: surface mesh
@@ -78,3 +78,105 @@ def test_SmartLaplacianSmoothing(M, options):
     assert M.NNode == Mnew.NNode, 'Number of nodes incorrectly altered.'
     assert M.NElem == Mnew.NElem, 'Number of elements incorrectly altered.'
 
+@pytest.mark.parametrize("M, h, labels", [
+    # Case 1: solid sphere
+    (
+        implicit.TetMesh(implicit.sphere([0,0,0], 1), [-1,1,-1,1,-1,1], .1),
+        .2,
+        None
+    ),
+    # Case 2: surface sphere
+    (
+        implicit.SurfaceMesh(implicit.sphere([0,0,0], 1), [-1,1,-1,1,-1,1], .1),
+        .2,
+        None
+    ),
+    # Case 3: labeled solid spgere
+    (
+        implicit.TetMesh(implicit.sphere([0,0,0], 1), [-1,1,-1,1,-1,1], .1),
+        .2,
+        True
+    ),
+    # Case 4: labeled surface sphere
+    (
+        implicit.SurfaceMesh(implicit.sphere([0,0,0], 1), [-1,1,-1,1,-1,1], .1),
+        .2,
+        True
+    ),
+    ])
+def test_Contract(M, h, labels):
+    if labels is True:
+        labels = np.zeros(M.NElem)
+        labels[M.Centroids[:,0]<0] = 1
+    Mnew = improvement.Contract(M, h, labels=labels)
+    
+    if labels is not None:
+        assert len(Mnew.ElemData['labels']) == Mnew.NElem
+    if M.Type == 'vol':
+        # check for inversions
+        assert np.all(quality.Volume(*Mnew) > 0), 'Inverted elements'
+    elif M.Type == 'surf':
+        assert len(Mnew.BoundaryNodes) == 0, 'Exposed Edges'
+
+@pytest.mark.parametrize("M, h, labels", [
+    # Case 1: solid sphere
+    (
+        implicit.TetMesh(implicit.sphere([0,0,0], 1), [-1,1,-1,1,-1,1], .1),
+        .05,
+        None
+    ),
+    # Case 2: surface sphere
+    (
+        implicit.SurfaceMesh(implicit.sphere([0,0,0], 1), [-1,1,-1,1,-1,1], .1),
+        .05,
+        None
+    ),
+    # Case 3: labeled solid sphere
+    (
+        implicit.TetMesh(implicit.sphere([0,0,0], 1), [-1,1,-1,1,-1,1], .1),
+        .05,
+        True
+    ),
+    # Case 4: labeled surface sphere
+    (
+        implicit.SurfaceMesh(implicit.sphere([0,0,0], 1), [-1,1,-1,1,-1,1], .1),
+        .05,
+        True
+    ),
+    ])
+def test_Split(M, h, labels):
+    if labels is True:
+        labels = np.zeros(M.NElem)
+        labels[M.Centroids[:,0]<0] = 1
+    Mnew = improvement.Split(M, h, labels=labels)
+    
+    if labels is not None:
+        assert len(Mnew.ElemData['labels']) == Mnew.NElem
+    if M.Type == 'vol':
+        # check for inversions
+        assert np.all(quality.Volume(*Mnew) > 0), 'Inverted elements'
+    elif M.Type == 'surf':
+        assert len(Mnew.BoundaryNodes) == 0, 'Exposed Edges'
+
[…]
