### 2026-02-28 931ffbf
docs: Documentation updates, additions to theory guide, performance evals
