+++ src/mymesh/__init__.py
@@ -13,6 +13,7 @@ Objects
     .. currentmodule:: mymesh.mesh
 
     mesh
+    dmesh
 
 
 .. currentmodule:: mymesh
