+++ docs/source/theory/mesh.rst
@@ -1,0 +1,292 @@
+Mesh
+====
+
+Standard mesh
+-------------
+(See also: :class:`~mymesh.mesh.mesh`, :ref:`Working_with_the_mesh_class`)
+
+
+The :class:`~mymesh.mesh.mesh` class is robust and feature rich, supporting arbitrary, mixed-element meshes, with a variety of methods and on-demand properties.
+It's based around two data structures:
+
+1. ``NodeCoords`` - an array containing the coordinates of the nodes, and
+2. ``NodeConn`` - an array (or list of lists) containing indices of the nodes that connect to form elements (or the 'node connectivity')
+
+Dictionaries ``NodeData`` and ``ElemData`` can be used to store scalar or vector data alongside the nodes and elements.
+
+Dynamic mesh
+------------
+(See also: :class:`~mymesh.mesh.dmesh`)
+
+:class:`~mymesh.mesh.dmesh` is a specialized mesh data structure for rapid modifications to a mesh's connectivity, such as during coarsening (:func:`~mymesh.improvement.Contract`) or :mod:`~mymesh.delaunay` triangulation.
+It has three key features that enables rapid modifications: amortized :math:`\mathcal{O}(1)` insertion by doubling, swap removal, and a doubly-linked list to track node-element connectivity.
+
+.. note::
+
+    :class:`dmesh` only supports single element type meshes, and has only been fully tested on purely triangular and purely tetrahedral meshes.
+
+Amortized :math:`\mathcal{O}(1)` insertion
+^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
+
+In dynamically changing meshes, nodes and/or elements may be repeatedly added and/or removed.
+When performing many such operations, the cost of resizing the arrays storing node and element information becomes significant.
+To more efficiently handle insertions into arrays, the arrays are maintained with buffer space at the end, and a pointer (e.g. NNode, NElem) that tracks where the end of the meaningful data is. 
+Inserting a value into the next available space is a simple and :math:`\mathcal{O}(1)` operation.
+Whenever that buffer space runs out, the arrays will still need to be resized. 
+Since we don't know how big the array *needs* to be, we can just double the length of the arrays to create plenty of space relative to the size of the existing arrays. 
+
+.. graphviz::
+
+    digraph structs {
+        rankdir = "LR"
+        node [shape=record];
+        
+        node [shape=none]; // Use 'none' because the HTML table provides the shape
+
+        nodecoords [pos="0,0!", label=<
+            <TABLE BORDER="0" CELLBORDER="1" CELLSPACING="0" CELLPADDING="4">
+                <TR>
+                    <TD COLSPAN="3">NodeCoords</TD>
+                </TR>
+                <TR>
+                    <TD>x<sub>0</sub></TD>
+                    <TD>y<sub>0</sub></TD>
+                    <TD>z<sub>0</sub></TD>
+                </TR>
+                <TR>
+                    <TD>x<sub>1</sub></TD>
+                    <TD>y<sub>1</sub></TD>
+                    <TD>z<sub>1</sub></TD>
+                </TR>
+                <TR>
+                    <TD>x<sub>2</sub></TD>
+                    <TD>y<sub>2</sub></TD>
+                    <TD>z<sub>2</sub></TD>
+                </TR>
+                <TR>
+                    <TD>x<sub>3</sub></TD>
+                    <TD>y<sub>3</sub></TD>
[…]
