+++ docs/source/verify.rst
@@ -1,9 +1,0 @@
-Verification
-============
-
-Verification and performance evaluation of select operations.
-
-.. toctree::
-    :maxdepth: 2
-    
-    verification/tree
