### 2026-02-28 ca4f9ea
docs: Update and add new examples
