### 2026-02-28 c126273
docs: create contributing guidelines
