+++ CONTRIBUTING.md
@@ -1,0 +1,17 @@
+# MyMesh Contributing Guidelines
+
+The MyMesh community is grateful for contributions in the form of issues (e.g. bug reports, feature requests) or pull requests. 
+Before making a pull request, it's recommended you first raise an issue or contact the project manager directly (toj@bu.edu) to ensure any contributions are in-line with the current and future directions of MyMesh. 
+
+Any communications or contributions fall under the perview of the [code of conduct](CONTRIBUTOR_CODE_OF_CONDUCT.md).
+
+Any contributions will be published under the [MyMesh license](LICENSE.md), thus by making a pull request, you affirm that you are authorized to make such a contribution (e.g. you employer or institution doesn't have any claims to the copyright).
+
+At this time, we are not interested in any contributions that are written primarily or entirely by artificial intelligence.
+
+# Conventions
+## Versioning
+MyMesh utilizes the semantic versioning (“[SemVer](https://semver.org/)”) scheme, where version numbers are formatted as “MAJOR.MINOR.PATCH”.
+
+## Commits
+Commits follow the “[Conventional Commits](https://www.conventionalcommits.org/)” scheme to support changelog generation via [commitizen](https://commitizen-tools.github.io/commitizen/). 
