### 2026-03-01 0e15c8d
Update issue templates
