### 2026-03-01 #10003 @RichardLitt
@TimJosephson Does what I described above work for you?
