### 2026-03-01 #10003 @RichardLitt
@TimJosephson You will need to add a State of the Field section to your paper. Take a look at the new requirements, here: https://joss.readthedocs.io/en/latest/review_criteria.html.
