### 2026-03-01 #10003 @TimJosephson
@RichardLitt Yes that sounds good! I'm just about done preparing a response/revisions based on @jorgensd comments. I'll also make the required changes to the paper to add a State of the Field section. I should have these changes completed and posted sometime this week.
