### 2026-03-03 dc5d601
docs: paper draft update
