+++ paper/paper.bib
@@ -1,150 +1,165 @@
-@article{Josephson2024b,
-abstract = {Synthetic bone graft scaffolds aim to generate new bone tissue and alleviate the limitations of autografts and allografts. To meet that aim, it is essential to have a design approach able to generate scaffold architectures that will promote bone formation. Here, we present a topology-varying design optimization method, the “mixed-topology” approach, that generates new designs from a set of starting structures. This approach was used with objective functions focusing on improving the scaffold's local mechanical microenvironments to mechanobiologically promote bone formation within the scaffold and constraints to ensure manufacturability and achieve desired macroscale properties. The results demonstrate that this approach can successfully generate scaffold designs with improved microenvironments, taking into account different combinations of relevant stimuli and constraints.},
-author = {Josephson, Timothy O. and Morgan, Elise F.},
-doi = {10.1007/S10237-024-01880-0},
-file = {:home/tim/Nextcloud/Documents/Papers/2024 - Josephson, Morgan - Mechanobiological optimization of scaffolds for bone tissue engineering.pdf:pdf},
-isbn = {0123456789},
-issn = {1617-7940},
-journal = {Biomechanics and Modeling in Mechanobiology},
-keywords = {Biological and Medical Physics,Biomedical Engineering and Bioengineering,Biophysics,Theoretical and Applied Mechanics},
-month = {jul},
-pages = {1--18},
-publisher = {Springer},
-title = {{Mechanobiological optimization of scaffolds for bone tissue engineering}},
-url = {https://link.springer.com/article/10.1007/s10237-024-01880-0},
-year = {2024}
+@misc{trimesh,
+   author = {Dawson-Haggerty},
+   title = {trimesh},
+   url = {https://trimesh.org/},
+   year = {2019}
 }
 @book{Schroeder2006,
-author = {Schroeder, Will and Martin, Ken and Lorensen, Bill},
-isbn = {978-1-930934-19-1},
-publisher = {Kitware},
-title = {{The Visualization Toolkit (4th ed.)}},
-year = {2006}
+   author = {Will Schroeder and Ken Martin and Bill Lorensen},
+   isbn = {978-1-930934-19-1},
+   publisher = {Kitware},
+   title = {The Visualization Toolkit (4th ed.)},
+   year = {2006}
+}
+@book{cgal,
+   author = {The CGAL Project},
+   edition = {6.1},
+   publisher = {CGAL Editorial Board},
+   title = {CGAL User and Reference Manual},
+   url = {https://doc.cgal.org/6.1/Manual/packages.html},
+   year = {2025}
+}
+@article{Geuzaine2009,
+   abstract = {Gmsh is an open-source 3-D finite element grid generator with a build-in CAD engine and post-processor. Its design goal is to provide a fast, light and user-friendly meshing tool with parametric input and advanced visualization capabilities. This paper presents the overall philosophy, the main design choices and some of the original algorithms implemented in Gmsh. © 2009 John Wiley & Sons, Ltd.},
+   author = {Christophe Geuzaine and Jean François Remacle},
+   doi = {10.1002/nme.2579},
+   issn = {00295981},
+   issue = {11},
+   journal = {International Journal for Numerical Methods in Engineering},
+   keywords = {Computer-aided design,Finite element method,Mesh generation,Open-source software,Post-processing},
+   month = {9},
+   pages = {1309-1331},
+   publisher = {John Wiley \& Sons, Ltd},
+   title = {Gmsh: A 3-D finite element mesh generator with built-in pre- and post-processing facilities},
+   volume = {79},
[…]
