### 2026-03-03 751e418
fix(improvement.SmartLaplacianSmoothing): update and speed up smart laplacian smoothing
