+++ src/mymesh/improvement.py
@@ -472,7 +472,7 @@ def TangentialLaplacianSmoothing(M, options=dict()):
 
     return Mnew
 
-def SmartLaplacianSmoothing(M, target='mean', TangentialSurface=True, labels=None, options=dict()):
+def SmartLaplacianSmoothing(M, TangentialSurface=True, labels=None, options=dict()):
     """
     Performs smart Laplacian smoothing :cite:p:`Freitag1997a`, repositioning 
     each node to the center of its adjacent nodes only if doing so doesn't
@@ -482,14 +482,6 @@ def SmartLaplacianSmoothing(M, target='mean', TangentialSurface=True, labels=Non
     ----------
     M : mymesh.mesh
         Mesh object to smooth
-    target : str
-        Determines whether criteria for repositioning a node, by default 'mean'. 
-
-        'mean' - repositioning is allowed if the average quality of the connected
-        elements doesn't decrease.
-
-        'min' - repositioning is allowed if the minimum quality of the connected
-        elements doesn't decrease.
     TangentialSurface : bool, optional
         Option to use tangential Laplacian smoothing on the surface (and interfaces, 
         if labels are provided), by default True.
@@ -520,9 +512,6 @@ def SmartLaplacianSmoothing(M, target='mean', TangentialSurface=True, labels=Non
         FixSurf : bool
             If true, all nodes on the surface will be held in place and only 
             interior nodes will be smoothed, by default False.
-        qualityFunc : function
-            Function used for computing quality. It is assumed that a larger
-            number corresponds to higher quality, be default quality.MeanRatio
         InPlace : bool
             If True, the input mesh is modified directly, otherwise a new copy of the mesh
             is created, by default False.
@@ -586,7 +575,6 @@ def SmartLaplacianSmoothing(M, target='mean', TangentialSurface=True, labels=Non
                         FixFeatures = False,
                         FixSurf = False,
                         FixEdge = True,
-                        qualityFunc = quality.MeanRatio,
                         InPlace = False
                     )
 
@@ -595,13 +583,11 @@ def SmartLaplacianSmoothing(M, target='mean', TangentialSurface=True, labels=Non
     FixedNodes = SmoothOptions['FixedNodes']
     tolerance = SmoothOptions['tolerance']
     iterate = SmoothOptions['iterate']
-    qualityFunc = SmoothOptions['qualityFunc']
     maxIter = SmoothOptions['maxIter']
     method = SmoothOptions['method']
     InPlace = SmoothOptions['InPlace']
 
     # Initialize
-    lens = np.array([len(n) for n in NodeNeighbors])
     NodeConn = np.asarray(NodeConn)
     if TangentialSurface:
         SurfNodes = set([n for elem in SurfConn for n in elem])
@@ -615,94 +601,24 @@ def SmartLaplacianSmoothing(M, target='mean', TangentialSurface=True, labels=Non
         for i in EdgeNodes:
             NodeNeighbors[i] = EdgeNodeNeighbors[i]
 
-    q = qualityFunc(NodeCoords, NodeConn)
+    nneighbors = np.array([len(ns) for ns in M.NodeNeighbors])
+    neighbor_indices = np.insert(np.cumsum(nneighbors),0,0)
+    flat_neighbors = np.array([n for ns in M.NodeNeighbors for n in ns])
+
+    conn_indices = np.insert(np.cumsum([len(ns) for ns in M.ElemConn]),0,0)
[…]
