+++ src/mymesh/quality.py
@@ -520,23 +520,134 @@ def MeanRatio(NodeCoords,NodeConn,verbose=False):
 
     return q
 
+def ModifiedMeanRatio(NodeCoords, NodeConn, verbose=False):
+
+    NodeCoords = np.asarray(NodeCoords)
+    NodeConn = np.asarray(NodeConn)
+
+    D = np.empty((len(NodeConn),3,3))
+    D[:,0,:] = NodeCoords[NodeConn[:,1]] - NodeCoords[NodeConn[:,0]]
+    D[:,1,:] = NodeCoords[NodeConn[:,2]] - NodeCoords[NodeConn[:,0]]
+    D[:,2,:] = NodeCoords[NodeConn[:,3]] - NodeCoords[NodeConn[:,0]]
+
+    # W = np.array([
+    #     [1, 1/2, 1/2],
+    #     [0, np.sqrt(3)/2, np.sqrt(3)/6],
+    #     [0, 0, np.sqrt(2/3)]
+    # ])            
+    Winv = np.array([[ 1.        , -0.57735027, -0.40824829],
+                    [ 0.        ,  1.15470054, -0.40824829],
+                    [ 0.        ,  0.        ,  1.22474487]])
+    
+    # Matrix multiplication in a numba-friendly style
+    S = np.dot(D.reshape(-1,3), Winv).reshape(D.shape)
+
+    Sfrob = np.linalg.norm(S, ord='fro', axis=(1,2))
+    det = np.linalg.det(S)
+    eps = np.finfo(float).eps
+    delta = np.sqrt(eps*(eps - det.min())) if det.min() < eps else 0
+    h = 0.5 * (det + np.sqrt(det**2 + 4*delta**2))
+
+    q = 3*h**(2/3) / Sfrob**2
+
+    if verbose:
+        minq = min(q)
+        maxq = max(q)
+        meanq = np.mean(q)
+        print('------------------------------------------')
+        print(f'Minimum Mean Ratio: {minq:.3f} on Element {np.where(q==minq)[0][0]:.0f}')
+        print(f'Maximum Mean Ratio: {maxq:.3f} on Element {np.where(q==maxq)[0][0]:.0f}')
+        print(f'Mean Mean Ratio: {meanq:.3f}')
+        print('------------------------------------------')
+    return q
+
 @try_njit
-def _MeanRatio(NodeCoords, NodeConn):
+def _MeanRatio(NodeCoords, elem):
 
-    W = np.array([
-        [1, 1/2, 1/2],
-        [0, np.sqrt(3)/2, np.sqrt(3)/6],
-        [0, 0, np.sqrt(2/3)]
-    ])            
-    Winv = np.linalg.inv(W)
+    # W = np.array([
+    #     [1, 1/2, 1/2],
+    #     [0, np.sqrt(3)/2, np.sqrt(3)/6],
+    #     [0, 0, np.sqrt(2/3)]
+    # ])            
+    # W inverse
+    w00, w01, w02 = 1, -0.57735027, -0.40824829
+    w10, w11, w12 = 0,  1.15470054, -0.40824829
+    w20, w21, w22 = 0,  0,  1.22474487
 
-    q = np.zeros(len(NodeConn))
-    for i in range(len(NodeConn)):
-        D = NodeCoords[NodeConn[i,np.array([1,2,3])]] - NodeCoords[NodeConn[i,0]] 
-        S = D @ Winv
[…]
