### 2026-03-07 9cbd93b
fix(converter.im2voxel): fix bug in im2voxel
