### 2026-03-07 2af6e4e
docs: update demo_image-based
