### 2026-03-07 c830bfd
docs: update demo_contouring
