### 2026-03-08 a5be0ac
docs: update workflow dependencies
