### 2026-03-08 18d81fd
docs: fix code execution in performance section
