+++ dev/_downloads/0eddfcdb2b42d7496bcc29b0269d59d3/contour-1.py
@@ -45,3 +45,4 @@ Spvmc = grid.contour([threshold], img.flatten(order='F'), method='flying_edges')
 pvfe_time = time.time()-tic
 
 plt.bar(['mc\n(image)', 'mc', 'mc33', 'mt', 'pv_mc', 'pv_fe', 'voxel'], [mcimage_time, mc_time, mc33_time, mtimage_time, pvmc_time, pvfe_time, vox_time])
+plt.ylabel('Time (s)')
