### 2026-03-08 327d9ad
docs: debug performance>curvature
