### 2026-03-08 55a9c63
docs: debug performance>curvature
