### 2026-03-08 a17bbfd
docs: debug performance>curvature
