+++ .github/workflows/documentation.yml
@@ -131,7 +131,7 @@ jobs:
           jupyter_rfb \
           triangle \
           pyvista \
-          trimesh
+          trimesh[easy]
         # Install vispy last
         pip install vispy
     - name: Install package
@@ -176,7 +176,7 @@ jobs:
       shell: bash -l {0}
       run: |
         export DISPLAY=:99.0
-        sphinx-build -W -b html ./docs/source ./docs/build
+        sphinx-build -b html ./docs/source ./docs/build
     - name: Run tests
       shell: bash -l {0}
       run: |
