### 2026-03-09 9252b82
docs: paper revisions
