+++ paper/paper.bib
@@ -1,3 +1,40 @@
+@article{Mtchedlishvili2026,
+   author = {Mariami Mtchedlishvili},
+   journal = {University College Dublin},
+   month = {3},
+   title = {Personal Communication},
+   year = {2026}
+}
+@article{Lim2026,
+   author = {Matthew Lim},
+   journal = {Boston University},
+   month = {3},
+   title = {Personal Communication},
+   year = {2026}
+}
+@article{Smeets2022b,
+   abstract = {Nanomesh is a Python library that allows users to quickly and easily create 2D and 3D meshes directly from images of the object they wish to mesh. The automated workflow can preprocess and segment the picture to extract different regions and create conforming meshes of the objects. Analysis tools allow evaluating the quality of the resuting mesh and the detection of problematic regions. The resulting meshes can be exported to a variety of popular formats so that they can be used in finite element simulations. Nanomesh can be used as python library for example in Jupyter notebooks, or through dedicated online dashboards.},
+   author = {Stef Smeets and Nicolas Renaud and Lars J. Corbijn van Willenswaard},
+   doi = {10.21105/joss.04654},
+   issn = {2475-9066},
+   issue = {78},
+   journal = {Journal of Open Source Software},
+   month = {10},
+   pages = {4654},
+   publisher = {The Open Journal},
+   title = {Nanomesh: A Python workflow tool for generating meshes from image data},
+   volume = {7},
+   url = {https://joss.theoj.org/papers/10.21105/joss.04654},
+   year = {2022}
+}
+@misc{Schlomer2021,
+   author = {Schlömer Nico},
+   doi = {https://doi.org/10.5281/zenodo.5564818},
+   publisher = {Zenodo},
+   title = {pygalmesh: Python interface for CGAL's meshing tools},
+   url = {https://github.com/nschloe/pygalmesh},
+   year = {2021}
+}
 @misc{trimesh,
    author = {Dawson-Haggerty},
    title = {trimesh},
@@ -139,27 +176,4 @@
    volume = {83},
    url = {https://www.doi.org/10.1007/BF01313411},
    year = {1991}
-}
-@article{Smeets2022b,
-   abstract = {Nanomesh is a Python library that allows users to quickly and easily create 2D and 3D meshes directly from images of the object they wish to mesh. The automated workflow can preprocess and segment the picture to extract different regions and create conforming meshes of the objects. Analysis tools allow evaluating the quality of the resuting mesh and the detection of problematic regions. The resulting meshes can be exported to a variety of popular formats so that they can be used in finite element simulations. Nanomesh can be used as python library for example in Jupyter notebooks, or through dedicated online dashboards.},
-   author = {Stef Smeets and Nicolas Renaud and Lars J. Corbijn van Willenswaard},
-   doi = {10.21105/joss.04654},
-   issn = {2475-9066},
-   issue = {78},
-   journal = {Journal of Open Source Software},
-   month = {10},
-   pages = {4654},
-   publisher = {The Open Journal},
-   title = {Nanomesh: A Python workflow tool for generating meshes from image data},
-   volume = {7},
-   url = {https://joss.theoj.org/papers/10.21105/joss.04654},
-   year = {2022}
-}
-@misc{Schlomer2021,
-   author = {Schlömer Nico},
-   doi = {https://doi.org/10.5281/zenodo.5564818},
-   publisher = {Zenodo},
-   title = {pygalmesh: Python interface for CGAL's meshing tools},
-   url = {https://github.com/nschloe/pygalmesh},
-   year = {2021}
 }
