### 2026-03-09 59c84d3
docs: Update optional dependency guide
