### 2026-03-10 b3a6312
tests: switch to isclose for knn distance checks
