### 2026-03-10 2ea1947
fix(delaunay.Triangle): fix inputs to handle constraints more reliably
