### 2026-03-22 #10003 @TimJosephson
Just to check in/provide an update, I received comments/issues from @jorgensd and made updates to the paper and repository based on that feedback a couple weeks ago. I see the checklist filled out by @Ingvord with all but two boxes checked, but haven't received any comments/feedback on specific issues that might need to be addressed. I'm happy to address any further feedback/suggestions/concerns.
