### 2026-03-24 d2336d4
fix(improvement.SmartLaplacianSmoothing): fix bug for TangentialSurface=False
