### 2026-03-24 BU-SMBL/mymesh#6 @jorgensd
[FEATURE] Surface and volume markers
Relates to https://github.com/openjournals/joss-reviews/issues/10003

In many problems met in engineering, there is a requirement for surface and volume markers of various sub-domains and sub-surfaces.
Is this a feature planned for the software?

The feature is not required for acceptance of the software paper, but something that I would see as crucial for complex meshes. Especially when it comes to CSG (https://bu-smbl.github.io/mymesh/dev/guide/csg.html)
