### 2026-03-24 #10003 @jorgensd
> Thank you. [@jorgensd](https://github.com/jorgensd): What do you think? [@Ingvord](https://github.com/Ingvord): Do you have specific comments about your remaining empty checkboxes?

@RichardLitt

I've raised some more issues, mainly relating to development practices.
I'm also not sure how I can justify ticking off the box:
> Collaborative effort: Does the project show evidence of community engagement such as contributions from multiple developers, issues/discussions from external users, or community influence on the development described in the paper?

As there is no online evidence of this being a collaborative project, as no PRs or issues has been made by anyone but me. Furthermore, the only contributor to code is the main author.
