### 2026-03-25 #10003 @RichardLitt
Thank you, @jorgensd. Looking at the [editorial guidelines](https://joss.readthedocs.io/en/latest/editing.html), there is room for the possibility of community being checked off as long as there are clear guidelines for others to commit. However, research impact also needs to be taken into account.

@TimJosephson Can you add any context for how this work has been used outside of your personal work?
