### 2026-04-03 806a2f2
feat(mesh.dmesh): add NodeLabels, additional methods
