+++ src/mymesh/mesh.py
@@ -9,7 +9,7 @@ mesh class
 
 """
 
-from . import utils, implicit, improvement, contour, converter, quality, rays, register, curvature, visualize, check_numba, try_njit
+from . import utils, implicit, improvement, contour, converter, quality, rays, register, curvature, visualize, delaunay, check_numba, try_njit
 from sys import getsizeof
 import scipy
 import numpy as np
@@ -2555,7 +2555,7 @@ class mesh:
         else:
             return [], [], []
 
-    def mesh2dmesh(self, ElemLabels=None):
+    def mesh2dmesh(self, NodeLabels=None, ElemLabels=None):
         """
         Convert to a dynamic mesh (:class:`dmesh`)
 
@@ -2577,12 +2577,16 @@ class mesh:
             ElemLabels = np.empty(0, np.int64)
         else:
             ElemLabels = np.asarray(ElemLabels, np.int64)
+        if NodeLabels is None:
+            NodeLabels = np.empty(0, np.int64)
+        else:
+            NodeLabels = np.asarray(NodeLabels, np.int64)
         if self.verbose:
             self.verbose = False
             _ = self.ElemConn
             self.verbose = True
 
-        D = dmesh(NodeCoords, NodeConn, self.ElemConn, ElemLabels)
+        D = dmesh(NodeCoords, NodeConn, self.ElemConn, ElemLabels, NodeLabels)
 
         return D
     
@@ -2627,10 +2631,18 @@ def dmesh_get_ElemConn_size(self):
 @try_njit(cache=True)
 def dmesh_get_ElemLabels(self):
     return self.ElemLabels
-
+@try_njit(cache=True)
+def dmesh_set_ElemLabels(self, labels):
+    self.ElemLabels = labels
+@try_njit(cache=True)
+def dmesh_get_NodeLabels(self):
+    return self.NodeLabels
+@try_njit(cache=True)
+def dmesh_set_NodeLabels(self, labels):
+    self.NodeLabels = labels
 # Method implementations:
 @try_njit(cache=True)
-def dmesh_addNodes(self,NodeCoords):
+def dmesh_addNodes(self,NodeCoords,NodeLabels=None):
     """
     Add nodes to the mesh.
 
@@ -2650,6 +2662,11 @@ def dmesh_addNodes(self,NodeCoords):
         # to exactly fit the new nodes
         newsize = np.maximum(NewLength,len(self.raw_NodeCoords)*2)
         self.raw_NodeCoords = np.resize(self.raw_NodeCoords, (newsize,3))
+        if len(self.NodeLabels) > 0:
+            self.NodeLabels = np.resize(self.NodeLabels, int(np.maximum(NewLength,len(self.NodeLabels)*2)))
+            if NodeLabels is None:
+                NodeLabels = np.zeros(len(NodeCoords), dtype=np.int64)
+            self.NodeLabels[self.NNode:NewLength] = NodeLabels
         
         # update the ElemConn structure as well
         self.ElemConn_head = np.resize(self.ElemConn_head, newsize)
@@ -2815,7 +2832,103 @@ def dmesh_removeElems(self, ElemIds):
     for ElemId in np.sort(ElemIds)[::-1]:
         dmesh_removeElem(self, ElemId)
 
-def dmesh_init(raw_NodeCoords, raw_NodeConn, ElemConn, ElemLabels=None):
+@try_njit
+def dmesh_get_TriEdgeNeighbor(self, tri_id, n1, n2):
+    # get the element that neighbors `tri_id` across edge (n1, n2)
+    next_tri_id = -1
+    i = self.ElemConn_head[n1]
+    while i != -1:
+        next_elem = self.ElemConn_elem[i]
+        # iterate through elem conn for the first node in the edge
+        if next_elem != tri_id:
[…]
