### 2026-04-03 a7bd8b4
feat(delaunay): Add support for 2D constrained Delaunay triangulation
