+++ src/mymesh/delaunay.py
@@ -65,8 +65,7 @@ def Triangulate(NodeCoords,Constraints=None,method=None,tol=1e-8, steiner=0):
         by default None. Edge constraints should be specified by node indices,
         for example [[0, 1], [1,2], ...]
     method : str, optional
-        Triangulation method, by default 'BowyerWatson' if no constraints are 
-        provided and 'Triangle' if constraints are provided.
+        Triangulation method, by default 'BowyerWatson'.
 
         - 'BowyerWatson' - Generate a Delaunay triangulation by the Bowyer-Watson algorithm (:func:`BowyerWatson2d`)
 
@@ -85,7 +84,7 @@ def Triangulate(NodeCoords,Constraints=None,method=None,tol=1e-8, steiner=0):
         if (Constraints is None or len(Constraints) == 0):
             method = 'BowyerWatson'
         else:
-            method = 'Triangle'
+            method = 'BowyerWatson'
 
     if (Constraints is None or len(Constraints) == 0):
         Points,_,idx = utils.DeleteDuplicateNodes(Points,[],return_idx=True, tol=tol)
@@ -97,16 +96,16 @@ def Triangulate(NodeCoords,Constraints=None,method=None,tol=1e-8, steiner=0):
             NodeCoords, NodeConn = Triangle(Points,steiner=steiner)
             NodeConn = idx[NodeConn]
         else:
-            raise ValueError(f'Invalid method "{method:s}".')
+            raise ValueError(f'Method "{method}" not supported for triangulation')
     else: 
         # Constrained Delaunay Triangulation - Sloan (1993)
         # Generate initial triangulation
-        if method.lower() == 'triangle':
-            method = 'triangle'
+        if method.lower() == 'bowyerwatson':
+            NodeConn = BowyerWatson2d(Points, Constraints=Constraints)
+        elif method.lower() == 'triangle':
+            NodeCoords, NodeConn = Triangle(Points,Constraints=Constraints,steiner=steiner)
         else:
-            raise ValueError('Currently only method="Triangle" is supported for constrained triangulation.')
-        
-        NodeCoords, NodeConn = Triangle(Points,Constraints=Constraints,steiner=steiner)
+            raise ValueError(f'Method "{method}" not supported for constrained triangulation')
 
     if 'mesh' in dir(mesh):
         T = mesh.mesh(NodeCoords,NodeConn)
@@ -493,7 +492,7 @@ def TriangleSplitting(NodeCoords, Hull=None):
 
     return NodeConn
         
-def BowyerWatson2d(NodeCoords, nsample=3):
+def BowyerWatson2d(NodeCoords, Constraints=None):
     """
     Bowyer-Watson algorithm for 2D Delaunay triangulation
 
@@ -550,15 +549,27 @@ def BowyerWatson2d(NodeCoords, nsample=3):
     
     d = m.mesh2dmesh()
 
-    d = _bowyer_watson_loop_2d(d, indices, nsample)
+    d = _bowyer_watson_loop_2d(d, indices)
 
-    NodeConn = d.NodeConn
-    Super = np.any(NodeConn == nPts, axis=1) | \
-        np.any(NodeConn == (nPts+1), axis=1) | \
-        np.any(NodeConn == (nPts+2), axis=1)
-    NodeConn = NodeConn[~Super]
+    # Remove super triangle
+    d.removeElems(d.getElemConn(nPts+2))
+    d.removeElems(d.getElemConn(nPts+1))
+    d.removeElems(d.getElemConn(nPts))
+    # Insert constraints
+    if Constraints is not None:
+        d.NodeLabels = np.zeros(len(d.raw_NodeCoords), dtype=np.int64)
[…]
