### 2026-04-03 05b408d
feat(delaunay.ConvexHull): revamp GiftWrapping and add QuickHull for 2D convex hulls
