+++ src/mymesh/delaunay.py
@@ -35,12 +35,12 @@ Hulls
     :toctree: submodules/
 
     ConvexHull
+    GiftWrapping2d
+    QuickHull2d
     AlphaShape
     Alpha2d
     Alpha3d
-    AlphaPeel3d
-    GiftWrapping
-    
+    AlphaPeel3d    
 
 """
 #%%
@@ -159,7 +159,7 @@ def Tetrahedralize(NodeCoords, method=None, tol=1e-8):
         T = mesh(NodeCoords,NodeConn)
     return T
 
-def ConvexHull(NodeCoords,method='scipy',OrientSurf=True):
+def ConvexHull(NodeCoords, method=None, OrientSurf=True, nD=None):
     """
     Identify the convex hull of a set of points. For a 2D point set 
     (np.shape(NodeCoords) = (n,2)), a 2D convex hull of line elements will be 
@@ -170,19 +170,21 @@ def ConvexHull(NodeCoords,method='scipy',OrientSurf=True):
     ----------
     NodeCoords : array_like
         Coordinates of points around which the convex hull will be identified.
-        If shape = (n,2), a 2D convex hull will be identified, or if shape = (n,3)
-        a 3D convex hull will be identified. 
     method : str, optional
-        Convex hull method, by default 'scipy'.
+        Convex hull method, by default 'QuickHull' for 2D and 'scipy' for 3D.
 
-        - 'scipy' - Use :external+scipy:class:`scipy.spatial.ConvexHull`
+        - 'QuickHull' - Use the quickhull algorithm (:func:`QuickHull2d`)
+        
+        - 'GiftWrapping' - Use the gift wrapping algorithm (:func:`GiftWrapping2d`)
+        
+        - 'scipy' or 'qhull' - Use qhull via :external+scipy:class:`scipy.spatial.ConvexHull`
 
-        - 'BowyerWatson' - Generate a Delaunay triangulation by the Bowyer-Watson algorithm (:func:`BowyerWatson2d` or :func:`BowyerWatson3d`)
+        - 'BowyerWatson' - Extract the boundary of a Delaunay triangulation by the Bowyer-Watson algorithm (:func:`BowyerWatson2d` or :func:`BowyerWatson3d`). This option is mostly for theoretical interest.
 
-        - 'GiftWrapping' - Use the gift wrapping algorithm (:func:`GiftWrapping`)
+        
     OrientSurf : str, optional
         Ensure the normals of the convex hull are consistently oriented outward,
-        by default True.
+        by default True. This is only relevant for method=='scipy', other methods are oriented regardless.
 
     Returns
     -------
@@ -191,13 +193,26 @@ def ConvexHull(NodeCoords,method='scipy',OrientSurf=True):
         or :code:`Hull.Type='surf'` for a 3D hull.
 
     """    
-    nD = np.shape(NodeCoords)[1]
+    if nD is None:
+        if np.shape(NodeCoords)[1] == 2:
+            nD = 2
+        elif np.all(NodeCoords[:,2] == 0):
+            nD = 2
+        elif np.shape(NodeCoords)[1] == 3:
+            nD = 3
+        else:
+            raise ValueError("n>3 dimensional convex hulls aren't supported, use scipy.spatial.ConvexHull directly.")
+
 
     if nD == 2:
-        if method.lower() == 'giftwrapping':
-            hull = GiftWrapping(NodeCoords,IncludeCollinear=True)
-        elif method.lower() == 'scipy':
-            qhull = spatial.ConvexHull(np.asarray(NodeCoords, dtype=np.float64))
+        if method is None:
+            method = 'quickhull'
+        if method.lower() == 'quickhull':
[…]
