### 2026-04-03 caff31f
docs: update delaunay theory and performance guides
