+++ docs/source/verification/delaunay.rst
@@ -5,15 +5,19 @@ Delaunay triangulation is a well studied problem, and there are several well-est
 
 The current implementations in :mod:`mymesh` don't yet meet or exceed the performance or functionality of other available options (though this is an area of active development), so using well established implementations is recommended:
 
-- :code:`qhull` (available via :func:`scipy.spatial.Delaunay`, or :func:`mymesh.delaunay.SciPy`) for n dimensional Delaunay triangulations
+- :code:`qhull` (available via :class:`scipy.spatial.Delaunay`, or :func:`mymesh.delaunay.SciPy`) for n dimensional Delaunay triangulations
 - :code:`triangle` for 2 dimensional constrained/unconstrained Delaunay triangulations
 - :code:`TetGen` for 3 dimensional constrained/unconstrained Delaunay triangulations
 - :code:`gmsh`
 - :code:`CGAL`
 
-Here you can see a comparison between the performance of :mod:`mymesh` and some alternatives 
+Here you can see a comparison between the performance of :mod:`mymesh` and some alternatives for Delaunay triangulation and other related methods.
 
 See also: :ref:`Delaunay Triangulation`, :mod:`mymesh.delaunay`
+
+
+2D Delaunay Triangulation
+-------------------------
 
 .. plot::
     :include-source: False
@@ -65,8 +69,6 @@ See also: :ref:`Delaunay Triangulation`, :mod:`mymesh.delaunay`
 
     N = 3
     for i,n in enumerate(npoints):
-        points = np.random.rand(n, 2)
-
         # BowyerWatson
         if i < 2 or bowyer_watson_avg[-1] < 5:
             reps = []
@@ -124,5 +126,229 @@ See also: :ref:`Delaunay Triangulation`, :mod:`mymesh.delaunay`
     plt.xlabel('# of points')
     plt.ylabel('Time (s)')
     plt.title('2D Delaunay Triangulation')
+    plt.grid()
+    plt.show()
+
+2D Constrained Delaunay Triangulation
+-------------------------------------
+
+.. plot::
+    :include-source: False
+    :show-source-link: True
+
+    import sys, os, time, subprocess
+    import scipy
+    import numpy as np
+    import matplotlib.pyplot as plt
+
+    def run_no_numba(npoints, seed):
+        code = f"""
+    import os, subprocess, time
+    import numpy as np
+    os.environ["NUMBA_DISABLE_JIT"] = "1"
+    from mymesh import delaunay
+    # pre-run in case of any first time delays
+    delaunay.BowyerWatson2d(np.random.rand(10,2), Constraints=np.random.choice(np.arange(0, 10, dtype=int), size=(3,2), replace=False))
+
+    np.random.seed({seed})
+    points = np.random.rand({npoints}, 2)
+    constraints = np.random.choice(np.arange(0, len(points), dtype=int), size=(3,2),replace=False)
+    tic = time.time()
+    delaunay.BowyerWatson2d(points, Constraints=constraints)
+    t = time.time()-tic
+    print(t)
+    """
+        result = subprocess.run([sys.executable, "-c", code], capture_output=True)
+        t = float(result.stdout)
+
+        return t
+
+
+    from mymesh import delaunay
+    npoints = np.array([1e1, 1e2, 1e3, 1e4, 1e5, 1e6, 1e7], dtype=int)
+
+    bowyer_watson_avg = []
+    bowyer_watson_std = []
+    bowyer_watson_no_numba_avg = []
+    bowyer_watson_no_numba_std = []
+    triangle_avg = []
+    triangle_std = []
+
+    # pre-run for any first time pre-compiling
+    points = np.random.rand(10,2)
[…]
