+++ docs/source/theory/delaunay.rst
@@ -59,32 +59,6 @@ of points with radius being the largest distance from the center to a point
         s3 -- s1 [penwidth=1, color="#d08770"]; 
     }
 
-The data structure for constructing the triangulation consists of two tables 
-(:code:`dict`) - an element table and an edge table. The element table, keyed by
-the node numbers that define the element, contain the oriented "half-edges" 
-connected to each element and the edge table is keyed by edges and contains 
-the element key connected to that edge. Each half-edge is connected to only one 
-element, it's "twin" has opposite numbering and is connected to a neighboring
-element, providing a structure allowing for efficient mesh traversal.
-
-The initial data structure for the super-triangle with vertices 
-(:math:`p_{1^*}`, :math:`p_{2^*}`, :math:`p_{3^*}`) is:
-
-+-----------------------------------------------------+----------------------------------------------------------------------------------------------------------------+
-| Elements                                            | Edges                                                                                                          |
-+=====================================================+================================================================================================================+
-| (:math:`p_{1^*}`, :math:`p_{2^*}`, :math:`p_{3^*}`) | (:math:`p_{1^*}`, :math:`p_{2^*}`), (:math:`p_{2^*}`, :math:`p_{3^*}`), (:math:`p_{3^*}`, :math:`p_{1^*}`)     |
-+-----------------------------------------------------+----------------------------------------------------------------------------------------------------------------+
-
-+-------------------------------------+-------------------------------------------------------+
-| Edges                               | Element                                               |
-+=====================================+=======================================================+
-| (:math:`p_{1^*}`, :math:`p_{2^*}`)  | (:math:`p_{1^*}`, :math:`p_{2^*}`, :math:`p_{3^*}`)   |
-+-------------------------------------+-------------------------------------------------------+
-| (:math:`p_{2^*}`, :math:`p_{3^*}`)  | (:math:`p_{1^*}`, :math:`p_{2^*}`, :math:`p_{3^*}`)   |
-+-------------------------------------+-------------------------------------------------------+
-| (:math:`p_{3^*}`, :math:`p_{2^*}`)  | (:math:`p_{1^*}`, :math:`p_{2^*}`, :math:`p_{3^*}`)   |
-+-------------------------------------+-------------------------------------------------------+
 
 Bowyer-Watson: Point Insertion
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
@@ -508,10 +482,7 @@ Retriangulation
 
 Once the cavity has been formed and the edges at the boundary of the cavity 
 identified, the vertices of the cavity can be simply connected to the inserted
-point to retriangulate the cavity. Due to the oriented nature of the half-edges
-in the data structure, the triangles can be formed in a way that ensures that 
-the points of every triangle are ordered counter clockwise. 
-
+point to retriangulate the cavity. 
 
 
 .. grid:: 2
@@ -588,4 +559,518 @@ the points of every triangle are ordered counter clockwise.
                 c -- f
                 d -- f 
[…]
