### 2026-04-03 66794ef
docs: update references.bib
