+++ docs/source/references.bib
@@ -1,3 +1,82 @@
+@article{Barber1996,
+   abstract = {<p>The convex hull of a set of points is the smallest convex set that contains the points. This article presents a practical convex hull algorithm that combines the two-dimensional Quickhull algorithm with the general-dimension Beneath-Beyond Algorithm. It is similar to the randomized, incremental algorithms for convex hull and delaunay triangulation. We provide empirical evidence that the algorithm runs faster when the input contains nonextreme points and that it used less memory. computational geometry algorithms have traditionally assumed that input sets are well behaved. When an algorithm is implemented with floating-point arithmetic, this assumption can lead to serous errors. We briefly describe a solution to this problem when computing the convex hull in two, three, or four dimensions. The output is a set of “thick” facets that contain all possible exact convex hulls of the input. A variation is effective in five or more dimensions.</p>},
+   author = {C. Bradford Barber and David P. Dobkin and Hannu Huhdanpaa},
+   doi = {10.1145/235815.235821},
+   issn = {0098-3500},
+   issue = {4},
+   journal = {ACM Transactions on Mathematical Software},
+   keywords = {Delaunay triangulation,I35 [Computer Graphics]: Computational Geometry and Object Modeling-geometric algorithms,Reliability Additional Key Words and Phrases: Convex hull,Voronoi diagram,and systems General Terms: Algorithms,halfspace intersec-tion,languages},
+   month = {12},
+   pages = {469-483},
+   title = {The quickhull algorithm for convex hulls},
+   volume = {22},
+   url = {https://dl.acm.org/doi/10.1145/235815.235821},
+   year = {1996}
+}
+@article{Mtchedlishvili2026,
+   author = {Mariami Mtchedlishvili},
+   journal = {University College Dublin},
+   month = {3},
+   title = {Personal Communication},
+   year = {2026}
+}
+@article{Lim2026,
+   author = {Matthew Lim},
+   journal = {Boston University},
+   month = {3},
+   title = {Personal Communication},
+   year = {2026}
+}
+@article{Smeets2022a,
+   abstract = {Nanomesh is a Python library that allows users to quickly and easily create 2D and 3D meshes directly from images of the object they wish to mesh. The automated workflow can preprocess and segment the picture to extract different regions and create conforming meshes of the objects. Analysis tools allow evaluating the quality of the resuting mesh and the detection of problematic regions. The resulting meshes can be exported to a variety of popular formats so that they can be used in finite element simulations. Nanomesh can be used as python library for example in Jupyter notebooks, or through dedicated online dashboards.},
+   author = {Stef Smeets and Nicolas Renaud and Lars J. Corbijn van Willenswaard},
+   doi = {10.21105/joss.04654},
+   issn = {2475-9066},
+   issue = {78},
+   journal = {Journal of Open Source Software},
+   month = {10},
+   pages = {4654},
+   publisher = {The Open Journal},
+   title = {Nanomesh: A Python workflow tool for generating meshes from image data},
+   volume = {7},
+   url = {https://joss.theoj.org/papers/10.21105/joss.04654},
+   year = {2022}
+}
+@article{Smeets2022b,
+   abstract = {Nanomesh is a Python library that allows users to quickly and easily create 2D and 3D meshes directly from images of the object they wish to mesh. The automated workflow can preprocess and segment the picture to extract different regions and create conforming meshes of the objects. Analysis tools allow evaluating the quality of the resuting mesh and the detection of problematic regions. The resulting meshes can be exported to a variety of popular formats so that they can be used in finite element simulations. Nanomesh can be used as python library for example in Jupyter notebooks, or through dedicated online dashboards.},
[…]
