### 2026-04-03 6621acd
tests(delaunay): add QuickHull to ConvexHull test
