### 2026-04-03 2ab1a76
fix(mesh): fix labels in mesh2dmesh conversion
