### 2026-04-04 c11fb76
docs: add system info to performance evals
