### 2026-04-04 2cbea39
docs: update dependencies for doc build
