+++ pyproject.toml
@@ -64,6 +64,16 @@ all = [
 test = [
     "pytest>=8.3.4",
     "trimesh>=4.11.2",
+    "sphinx>=8.1.3",
+    "pydata-sphinx-theme>=0.16.1",
+    "sphinx-gallery>=0.18.0",
+    "sphinxcontrib-bibtex>=2.6.3",
+    "sphinx-copybutton>=0.5.2",
+    "jupyter-sphinx>=0.5.3",
+    "sphinx-design>=0.6.1",
+    "sphinx-tabs>=3.4.7",
+    "myst-parser>=4.0.1",
+    "py-cpuinfo>=9.0.0"
     ]
 
 [project.urls]
