### 2026-04-05 e39b4c2
docs: update rendering options in documentation build
