+++ docs/source/conf.py
@@ -112,7 +112,7 @@ import numpy as np
 from matplotlib import pyplot as plt
 import mymesh
 from mymesh import *
-visualize.set_vispy_backend(preference='PyQt6')
+visualize.set_vispy_backend(preference='osmesa')
 '''
 
 # graphviz_dot = r"C:\Program Files\Graphviz\bin\neato.exe"
