+++ docs/clean.sh
@@ -1,0 +1,8 @@
+#!/bin/bash
+# Deletes all files generated by the build
+shopt -s dotglob # find hidden directories that start with .
+docsdir="$(dirname "$0")"
+rm -rf "$docsdir/build"/*
+rm -rf "$docsdir/jupyter_execute"
+rm -rf "$docsdir/source/examples"/*
+rm -rf "$docsdir/source/generated"/*
