+++ docs/build.sh
@@ -1,0 +1,3 @@
+#!/bin/bash
+docsdir="$(dirname "$0")"
+sphinx-build -M html "$docsdir"/source/ "$docsdir"/build/
