+++ docs/README.md
@@ -6,6 +6,14 @@ pip install -U sphinx sphinx-design sphinx-copybutton sphinxcontrib-bibtex matpl
 ```
 
 ## Build:
+The documentation can be built with:
 ```
 sphinx-build -M html docs/source/ docs/build/
 ```
+or using `./docs/build.sh` (from the mymesh root directory).
+
+To test the documentation github workflow (./.github/workflows/documentation.yml) to verify the documentation builds correctly, use `./docs/act_build.sh` to build the docs with [act](https://github.com/nektos/act).
+
+To remove all files generated by the build process, run `./docs/clean.sh`.
+
+
