### 2026-04-05 #10003 @TimJosephson
Thanks @RichardLitt, for some additional context, mymesh has been used by a number of different researchers on projects beyond my own research work, as noted in the paper. There are multiple research publications utilizing mymesh currently in preparation or review, by myself, by others within our research group, and by researchers in other groups. Many of these users have made collaborative contributions to mymesh as they have found new ways to use it, reported bugs and unexpected behavior, and requested new features. However, since this occurred primarily within a research group and through personal communication, rather than an 'open source software' community, there isn't the same visibility of that collaboration. As mentioned in response to @Jorgensd's [original comment](https://github.com/BU-SMBL/mymesh/issues/2), future users who raise issues with me directly will be encouraged to make public issues and/or contributions on Github, and we updated contribution guidelines and issue templates to facilitate further contributions.

If you believe this needs to be further explained in the paper to meet the reviewer/editorial guidelines, we'd be happy to make additional clarifications.
