### 2026-04-05 BU-SMBL/mymesh#6 @TimJosephson
It depends on what exactly you mean by "markers", but `mymesh` has a couple mechanisms to achieve this (or something similar). The `mesh` object has dictionary properties NodeData/ElemData and NodeSets/ElemSets. NodeData/ElemData can be used to store various node- or element- associated data, such as scalar/vector fields or label arrays (e.g. can be used as shown in the [coarsening example](https://bu-smbl.github.io/mymesh/dev/examples/MeshModification/demo_coarsening.html#label-preserving-coarsening)). These structures are described [here](https://bu-smbl.github.io/mymesh/dev/guide/mesh_class.html#storing-data-with-meshes). NodeSets/ElemSets are similar but can be used to store sets of nodes/elements that belong to a specific region/group. The `Set` dictionaries aren't currently as integrated into other mymesh functionality as the `Data`, dictionaries but are available for users to store information. If you think there's a need that these structures don't meet I'd be happy to explore ways to meet that need.
