### 2026-04-22 #10003 @RichardLitt
Thank you, @TimJosephson. I am satisfied with that answer. If any papers have been published or are in review currently, it would be nice to cite them in the paper.

@TimJosephson Have you addressed the other issues that @jorgensd has raised?
