### 2026-04-24 0e92e88
docs(paper): Add Olowu et al (in review) reference
