+++ paper/paper.bib
@@ -1,3 +1,9 @@
+@article{Olowu2026,
+   author = {Benajmin D. Olowu and Joshua D. Auger and Elise F. Morgan},
+   journal = {Bone (In Review)},
+   title = {Validation of MRI- and CT-based Finite Element Modeling of the Proximal Femur Under Sideways Fall},
+   year = {2026}
+}
 @article{Mtchedlishvili2026,
    author = {Mariami Mtchedlishvili},
    journal = {University College Dublin},
