### 2026-04-24 #10003 @TimJosephson
That's great @RichardLitt, thank you. There's one paper currently in review that I've added to the paper.

I've responded to the issues @jorgensd raised, but I'm not sure if he has any follow up comments/questions.
