### 2026-04-29 BU-SMBL/mymesh#6 @jorgensd
So what I am I'm thinking about is for instance a multi-material simulation, where one would have trouble describing each material geometrically post mesh generation, such as:

from
https://scientificcomputing.github.io/fenics-in-the-wild/src/refinement_and_surface_markers.html
These volumes and surfaces has to be marked when one creates the mesh, not post-creation, as one does no longer have the information to backward engineer it.
