### 2026-04-29 #10003 @jorgensd
I think it is ok to accept this for publication, based on the following:
- A clear, educational documentation and user guide
- After revisions, a variety of examples is available: https://bu-smbl.github.io/mymesh/dev/examples/index.html#mesh-analysis-examples

My main concern, and why I've been hesitant to accept it, is that it currently only has a single developer, and very few users and no public community interaction, and thus looks like "yet another meshing tool". Let's hope that is not the case and that usage picks up post publication.

Additionally, I think how one defines performance varies. The performance section of the documentation relates to the mainly relates accuracy of the software, not the runtime (except: https://bu-smbl.github.io/mymesh/dev/verification/delaunay.html, but in my opinion it works on a rather small grid and doesn't compare to other meshing tools such as netgen, gmsh etc).
