### 2026-04-29 BU-SMBL/mymesh#8 @jorgensd
Joss paper references
Minor capitalization errors in JOSS paper:
https://github.com/openjournals/joss-reviews/issues/10003

Missing capitalization in:

```
Schnering, H. G. von, & Nesper, R. (1991). Nodal surfaces of fourier series: Fundamental
161 invariants of structured matter. Zeitschrift Für Physik B Condensed Matter, 83, 407–412.
162 https://doi.org/10.1007/BF01313411
```
should be

Schnering, H. G. von, & Nesper, R. (1991). Nodal surfaces of **F**ourier series: Fundamental invariants of structured matter. Zeitschrift Für Physik B Condensed Matter, 83, 407–412.
https://doi.org/10.1007/BF01313411

```
Shewchuk, J. R. (1996). Triangle: Engineering a 2D quality mesh generator and delaunay ....
```
should be

Shewchuk, J. R. (1996). Triangle: Engineering a 2D quality mesh generator and **D**elaunay ....

```
Smeets, S., Renaud, N., & Willenswaard, L. J. C. van. (2022). Nanomesh: A python workflow
```
should be

Smeets, S., Renaud, N., & Willenswaard, L. J. C. van. (2022). Nanomesh: A Python workflow
