### 2026-04-29 #10003 @RichardLitt
Thanks @jorgensd.

@TimJosephson If you could fix the references that @jorgensd points out, and add a note about performance or respond to that comment, that would be great. I think we can then move forward.
