### 2026-04-29 9a450b1
docs: Update capitalization in paper.bib
