+++ paper/paper.bib
@@ -131,7 +131,7 @@
 		PUB27
 		New York, NY, USA
 	},
-   title = {TetGen, a delaunay-based quality tetrahedral mesh generator},
+   title = {TetGen, a Delaunay-based quality tetrahedral mesh generator},
    volume = {41},
    url = {http://dx.doi.org/10.1145/2629697 https://dl.acm.org/doi/10.1145/2629697},
    year = {2015}
@@ -144,7 +144,7 @@
    issn = {16113349},
    journal = {Workshop on applied computational geometry},
    pages = {203-222},
-   title = {Triangle: Engineering a 2D quality mesh generator and delaunay triangulator},
+   title = {Triangle: Engineering a 2D quality mesh generator and Delaunay triangulator},
    url = {http://www.cs.cmu.edu/quake/triangle.html},
    year = {1996}
 }
