### 2026-04-29 9f160e4
docs: update performance docs
