+++ docs/source/verification/contour.rst
@@ -1,59 +1,158 @@
 Contour
 =======
 
+Contouring methods are used to extract surfaces from scalar fields (e.g. images or implicit functions). 
+MyMesh has several different contouring methods with different trade-offs and capabilities.
+Here, the efficiency of these methods are compared against each other and against contouring methods provided in `PyVista <https://docs.pyvista.org/>`_ (VTK) :cite:`Sullivan2019`. 
+
+See also: :ref:`theory_contour`, :mod:`mymesh.contour`, :ref:`contouring`
+
+3D Surface Contouring
+---------------------
+
+Contouring of a unit sphere in a uniform grid
 
 .. plot::
-    :include-source: True
+    :include-source: False
     :show-source-link: True
 
-    import time
+    import sys, os, time, subprocess
     import scipy
     import numpy as np
-    import mymesh
-    from mymesh import image, contour
-    import pyvista as pv
     import matplotlib.pyplot as plt
+    import pyvista as pv
 
-    img = mymesh.demo_image(scalefactor=0.25)
-    threshold = 100
-    h = (0.337891, 0.337891, 0.5)
+    from mymesh import primitives, implicit, contour
+    voxel_size = np.array([0.1, 0.1/2, 0.1/4, 0.1/8])#, 0.1/16])
 
-    tic = time.time()
-    Smcimage = image.SurfaceMesh(img, h, threshold, method='mc')
-    mcimage_time = time.time()-tic
-
-    tic = time.time()
-    Smc = image.SurfaceMesh(img, h, threshold, method='original')
-    mc_time = time.time()-tic
-
-    tic = time.time()
-    Smc33 = image.SurfaceMesh(img, h, threshold, method='mc33')
-    mc33_time = time.time()-tic
-
-    tic = time.time()
-    Smtimage = image.SurfaceMesh(img, h, threshold, method='mt')
-    mtimage_time = time.time()-tic
-
-    tic = time.time()
-    vox = image.VoxelMesh(img, h, threshold)
-    vox_time = time.time()-tic
-
-    grid = pv.ImageData(
+    mc_avg = []
+    mc_std = []
+    mc_img_avg = []
+    mc_img_std = []
+    mc33_avg = []
+    mc33_std = []
+    mt_avg = []
+    mt_std = []
+    pvmc_avg = []
+    pvmc_std = []
+    pvfe_avg = []
+    pvfe_std = []
+    # pre-run for any first time pre-compiling
+    h = 0.25
+    grid = primitives.Grid([-1,1,-1,1,-1,1], h)
+    grid.NodeData['f'] = implicit.sphere([0,0,0],1)(*grid.NodeCoords.T)
+    tetgrid = primitives.Grid([-1,1,-1,1,-1,1], h, ElemType='tet')
+    tetgrid.NodeData['f'] = implicit.wrapfunc(implicit.gyroid)(*tetgrid.NodeCoords.T)
+    img = grid.NodeData['f'].reshape(np.repeat(int(np.round((grid.NNode)**(1/3))),3), order='F')
+    pvgrid = pv.ImageData(
         dimensions=img.shape,
-        spacing=h,
+        spacing=(h,h,h),
         origin=(0,0,0),
     )
+    contour.MarchingCubes(grid.NodeCoords, grid.NodeConn, grid.NodeData['f'], method='original')
+    contour.MarchingCubesImage(img, method='original')
+    contour.MarchingCubes(grid.NodeCoords, grid.NodeConn, grid.NodeData['f'], method='33')
+    contour.MarchingTetrahedra(grid.NodeCoords, grid.NodeConn, grid.NodeData['f'])
+    pvgrid.contour([0], grid.NodeData['f'], method='marching_cubes') 
[…]
