+++ docs/source/references.bib
@@ -1,14 +1,33 @@
+@article{Olowu2026,
+   author = {Benajmin D. Olowu and Joshua D. Auger and Elise F. Morgan},
+   journal = {Bone (In Review)},
+   title = {Validation of MRI- and CT-based Finite Element Modeling of the Proximal Femur Under Sideways Fall},
+   year = {2026}
+}
+@inproceedings{Chew1993,
+   abstract = {For several commonly-used solution techniques for partial dhlerential equations, the first step is to divide the problem region into simply-shaped elements, creating a mesh. We present a technique for creating high-quality triangular meshes for regions on curved surfaces. This technique is an extension of previous methods we developed for regions in the plane.},
+   author = {L. Paul Chew},
+   city = {New York, New York, USA},
+   doi = {10.1145/160985.161150},
+   isbn = {0897915828},
+   booktitle = {Proceedings of the ninth annual symposium on Computational geometry  - SCG '93},
+   pages = {274-280},
+   publisher = {ACM Press},
+   title = {Guaranteed-quality mesh generation for curved surfaces},
+   url = {http://portal.acm.org/citation.cfm?doid=160985.161150},
+   year = {1993}
+}
 @article{Barber1996,
-   abstract = {<p>The convex hull of a set of points is the smallest convex set that contains the points. This article presents a practical convex hull algorithm that combines the two-dimensional Quickhull algorithm with the general-dimension Beneath-Beyond Algorithm. It is similar to the randomized, incremental algorithms for convex hull and delaunay triangulation. We provide empirical evidence that the algorithm runs faster when the input contains nonextreme points and that it used less memory. computational geometry algorithms have traditionally assumed that input sets are well behaved. When an algorithm is implemented with floating-point arithmetic, this assumption can lead to serous errors. We briefly describe a solution to this problem when computing the convex hull in two, three, or four dimensions. The output is a set of “thick” facets that contain all possible exact convex hulls of the input. A variation is effective in five or more dimensions.</p>},
+   abstract = {The convex hull of a set of points is the smallest convex set that contains the points. This article presents a practical convex hull algorithm that combines the two-dimensional Quickhull Algorithm with the general-dimension Beneath-Beyond Algorithm. It is similar to the randomized, incremental algorithms for convex hull and Delaunay triangulation. We provide empirical evidence that the algorithm runs faster when the input contains nonextreme points and that it uses less memory. Computational geometry algorithms have traditionally assumed that input sets are well behaved. When an algorithm is implemented with floating-point arithmetic, this assumption can lead to serious errors. We briefly describe a solution to this problem when computing the convex hull in two, three, or four dimensions. The output is a set of "thick" facets that contain all possible exact convex hulls of the input. A variation is effective in five or more dimensions.},
    author = {C. Bradford Barber and David P. Dobkin and Hannu Huhdanpaa},
    doi = {10.1145/235815.235821},
-   issn = {0098-3500},
+   issn = {00983500},
    issue = {4},
    journal = {ACM Transactions on Mathematical Software},
-   keywords = {Delaunay triangulation,I35 [Computer Graphics]: Computational Geometry and Object Modeling-geometric algorithms,Reliability Additional Key Words and Phrases: Convex hull,Voronoi diagram,and systems General Terms: Algorithms,halfspace intersec-tion,languages},
[…]
