### 2026-04-29 b7bc0a2
docs: Update capitalization in paper.bib
