+++ paper/paper.bib
@@ -28,7 +28,7 @@
    month = {10},
    pages = {4654},
    publisher = {The Open Journal},
-   title = {Nanomesh: A Python workflow tool for generating meshes from image data},
+   title = {Nanomesh: A {Python} workflow tool for generating meshes from image data},
    volume = {7},
    url = {https://joss.theoj.org/papers/10.21105/joss.04654},
    year = {2022}
@@ -131,7 +131,7 @@
 		PUB27
 		New York, NY, USA
 	},
-   title = {TetGen, a Delaunay-based quality tetrahedral mesh generator},
+   title = {TetGen, a {Delaunay}-based quality tetrahedral mesh generator},
    volume = {41},
    url = {http://dx.doi.org/10.1145/2629697 https://dl.acm.org/doi/10.1145/2629697},
    year = {2015}
@@ -144,7 +144,7 @@
    issn = {16113349},
    journal = {Workshop on applied computational geometry},
    pages = {203-222},
-   title = {Triangle: Engineering a 2D quality mesh generator and Delaunay triangulator},
+   title = {Triangle: Engineering a 2D quality mesh generator and {Delaunay} triangulator},
    url = {http://www.cs.cmu.edu/quake/triangle.html},
    year = {1996}
 }
@@ -178,7 +178,7 @@
    issue = {3},
    journal = {Zeitschrift für Physik B Condensed Matter},
    pages = {407-412},
-   title = {Nodal surfaces of Fourier series: Fundamental invariants of structured matter},
+   title = {Nodal surfaces of {Fourier} series: Fundamental invariants of structured matter},
    volume = {83},
    url = {https://www.doi.org/10.1007/BF01313411},
    year = {1991}
