### 2026-04-30 261eb97
docs: update plot colors for performance
