+++ dev/plot_directive/verification/contour-1.py
@@ -1,48 +1,134 @@
-import time
+import sys, os, time, subprocess
 import scipy
 import numpy as np
-import mymesh
-from mymesh import image, contour
-import pyvista as pv
 import matplotlib.pyplot as plt
+import pyvista as pv
 
-img = mymesh.demo_image(scalefactor=0.25)
-threshold = 100
-h = (0.337891, 0.337891, 0.5)
+from mymesh import primitives, implicit, contour
+voxel_size = np.array([0.1, 0.1/2, 0.1/4, 0.1/8])#, 0.1/16])
 
-tic = time.time()
-Smcimage = image.SurfaceMesh(img, h, threshold, method='mc')
-mcimage_time = time.time()-tic
-
-tic = time.time()
-Smc = image.SurfaceMesh(img, h, threshold, method='original')
-mc_time = time.time()-tic
-
-tic = time.time()
-Smc33 = image.SurfaceMesh(img, h, threshold, method='mc33')
-mc33_time = time.time()-tic
-
-tic = time.time()
-Smtimage = image.SurfaceMesh(img, h, threshold, method='mt')
-mtimage_time = time.time()-tic
-
-tic = time.time()
-vox = image.VoxelMesh(img, h, threshold)
-vox_time = time.time()-tic
-
-grid = pv.ImageData(
+mc_avg = []
+mc_std = []
+mc_img_avg = []
+mc_img_std = []
+mc33_avg = []
+mc33_std = []
+mt_avg = []
+mt_std = []
+pvmc_avg = []
+pvmc_std = []
+pvfe_avg = []
+pvfe_std = []
+# pre-run for any first time pre-compiling
+h = 0.25
+grid = primitives.Grid([-1,1,-1,1,-1,1], h)
+grid.NodeData['f'] = implicit.sphere([0,0,0],1)(*grid.NodeCoords.T)
+tetgrid = primitives.Grid([-1,1,-1,1,-1,1], h, ElemType='tet')
+tetgrid.NodeData['f'] = implicit.wrapfunc(implicit.gyroid)(*tetgrid.NodeCoords.T)
+img = grid.NodeData['f'].reshape(np.repeat(int(np.round((grid.NNode)**(1/3))),3), order='F')
+pvgrid = pv.ImageData(
     dimensions=img.shape,
-    spacing=h,
+    spacing=(h,h,h),
     origin=(0,0,0),
 )
+contour.MarchingCubes(grid.NodeCoords, grid.NodeConn, grid.NodeData['f'], method='original')
+contour.MarchingCubesImage(img, method='original')
+contour.MarchingCubes(grid.NodeCoords, grid.NodeConn, grid.NodeData['f'], method='33')
+contour.MarchingTetrahedra(grid.NodeCoords, grid.NodeConn, grid.NodeData['f'])
+pvgrid.contour([0], grid.NodeData['f'], method='marching_cubes')
+pvgrid.contour([0], grid.NodeData['f'], method='flying_edges')
 
-tic = time.time()
-Spvmc = grid.contour([threshold], img.flatten(order='F'), method='marching_cubes')
-pvmc_time = time.time()-tic
+N = 3
+npoints = []
+for i,h in enumerate(voxel_size):
 
-tic = time.time()
-Spvmc = grid.contour([threshold], img.flatten(order='F'), method='flying_edges')
-pvfe_time = time.time()-tic
+    grid = primitives.Grid([-1,1,-1,1,-1,1], h)
+    grid.NodeData['f'] = implicit.sphere([0,0,0],1)(*grid.NodeCoords.T)
+    tetgrid = primitives.Grid([-1,1,-1,1,-1,1], h, ElemType='tet')
[…]
