+++ dev/_downloads/cbb610dcfda8fd5b30583caae8b973dd/delaunay-2.py
@@ -82,9 +82,9 @@ for i,n in enumerate(npoints):
         triangle_std.append(np.std(reps))
 
 # Plot
-plt.errorbar(npoints[:len(bowyer_watson_avg)], bowyer_watson_avg, yerr=bowyer_watson_std, color='red')
-plt.errorbar(npoints[:len(bowyer_watson_no_numba_avg)], bowyer_watson_no_numba_avg, yerr=bowyer_watson_no_numba_std, color='pink')
-plt.errorbar(npoints[:len(triangle_avg)], triangle_avg, yerr=triangle_std, color='purple')
+plt.errorbar(npoints[:len(bowyer_watson_avg)], bowyer_watson_avg, yerr=bowyer_watson_std, color='#bf616a')
+plt.errorbar(npoints[:len(bowyer_watson_no_numba_avg)], bowyer_watson_no_numba_avg, yerr=bowyer_watson_no_numba_std, color='#5e81ac')
+plt.errorbar(npoints[:len(triangle_avg)], triangle_avg, yerr=triangle_std, color='#ebcb8b', linestyle='dotted')
 plt.xscale('log')
 plt.yscale('log')
 plt.legend(['mymesh', 'mymesh (no numba)', 'Triangle'])
