+++ 0.4.0/_sources/generated/submodules/mymesh.curvature.ImageCurvature.rst.txt
@@ -1,0 +1,6 @@
+mymesh.curvature.ImageCurvature
+===============================
+
+.. currentmodule:: mymesh.curvature
+
+.. autofunction:: ImageCurvature
