+++ 0.4.0/_sources/generated/submodules/mymesh.improvement.TangentialLaplacianSmoothing.rst.txt
@@ -1,0 +1,6 @@
+mymesh.improvement.TangentialLaplacianSmoothing
+===============================================
+
+.. currentmodule:: mymesh.improvement
+
+.. autofunction:: TangentialLaplacianSmoothing
