+++ 0.4.0/_sources/generated/submodules/mymesh.implicit.thicken.rst.txt
@@ -1,0 +1,6 @@
+mymesh.implicit.thicken
+=======================
+
+.. currentmodule:: mymesh.implicit
+
+.. autofunction:: thicken
