+++ 0.4.0/_sources/generated/submodules/mymesh.converter.tri2edges.rst.txt
@@ -1,0 +1,6 @@
+mymesh.converter.tri2edges
+==========================
+
+.. currentmodule:: mymesh.converter
+
+.. autofunction:: tri2edges
