+++ 0.4.0/_sources/generated/submodules/mymesh.implicit.VoxelMesh.rst.txt
@@ -1,0 +1,6 @@
+mymesh.implicit.VoxelMesh
+=========================
+
+.. currentmodule:: mymesh.implicit
+
+.. autofunction:: VoxelMesh

[…]
