### 2026-04-30 #10003 @TimJosephson
Thank you both @RichardLitt and @jorgensd !
I've fixed the references. My intent for the "performance" section is to have a combination of both accuracy/correctness and efficiency. It's currently limited to a subset of functionality, but includes the assessments of efficiency for several things, with comparisons to alternative softwares (scipy, triangle, and pyvista). I've just updated the efficiency evaluation for [contouring](https://bu-smbl.github.io/mymesh/dev/verification/contour.html), since that one was a bit sparse before. I intend to expand this section of the documentation further, including comparison to additional meshing tools, but I think what is currently included is sufficient to support the claims of the paper.

I've made a new release (0.4.0) that contains the changes made throughout the review process. Please let me know anything else you need from me to move forward!
