### 2026-04-30 BU-SMBL/mymesh#8 @TimJosephson
Thanks for catching that! I didn't realize the bibtex -> pdf generator enforced its own capitalization rules, so I needed to enforce capitalization with brackets in the bibtex file ("...{Fourier}..."). I had also missed the Delaunay capitalization in the bibtex. This should be resolved now.
