### 2026-04-30 BU-SMBL/mymesh#6 @TimJosephson
Ah I see what you mean. I think the properties I described (e.g. `mesh.ElemData`) are sufficient infrastructure to support this type of functionality, though I don't currently have functions overtly geared towards this type of mixed-element mesh.
I'll follow up with an example of how that type of mixed-element mesh could be made from an image with mymesh, and I'll think about other ways to better support this for other types of mesh generation
