### 2026-05-04 d7c56c7
docs: build paper pdf only for paper branch
