+++ .github/workflows/draft-pdf.yml
@@ -1,6 +1,8 @@
 name: Draft PDF
 on:
   push:
+    branches:
+        - paper
     paths:
       - paper/**
       - .github/workflows/draft-pdf.yml
