### 2026-05-19 d465756
docs(paper): Fix references
