+++ paper/paper.bib
@@ -1,7 +1,67 @@
+@misc{Lam2026,
+   author = {Siu Kwan Lam and stuartarchibald and Antoine Pitrou and Mark Florisson and Graham Markall and Stan Seibert and Emergency Self-Construct and Todd A Anderson and Guilherme Leobas and rjenc29 and Kaustubh and Swapnil Patel and Michael Collison and luk-f-a and Jay Bourque and Aaron Meurer and Travis E Oliphant and Michael Wang and densmirn and KrisMinchev and Andre Masella and Eric Wieser and njwhite and Ethan Pronovost and Ehsan Totoni and Stefan Seefeld and Hernan Grecco and Pearu Peterson and Isaac Virshup and MattyG},
+   doi = {10.5281/zenodo.19720971},
+   month = {4},
+   publisher = {Zenodo},
+   title = {numba},
+   url = {https://doi.org/10.5281/zenodo.19720971},
+   year = {2026}
+}
+@misc{Baratta2023,
+   author = {Igor A Baratta and Joseph P Dean and Jørgen S Dokken and Michal Habera and Jack S Hale and Chris N Richardson and Marie E Rognes and Matthew W Scroggs and Nathan Sime and Garth N Wells},
+   doi = {10.5281/zenodo.10447666},
+   month = {12},
+   publisher = {Zenodo},
+   title = {DOLFINx: The next generation FEniCS problem
+solving environment
+},
+   url = {https://doi.org/10.5281/zenodo.10447666},
+   year = {2023}
+}
+
+@misc{Schlomer2024,
+   author = {Nico Schlömer},
+   doi = {10.5281/zenodo.1173115},
+   title = {meshio: Tools for mesh files},
+   url = {https://github.com/nschloe/meshio},
+   year = {2024}
+}
+@article{Harris2006,
+   abstract = {<p> Array programming provides a powerful, compact and expressive syntax for accessing, manipulating and operating on data in vectors, matrices and higher-dimensional arrays. NumPy is the primary array programming library for the Python language. It has an essential role in research analysis pipelines in fields as diverse as physics, chemistry, astronomy, geoscience, biology, psychology, materials science, engineering, finance and economics. For example, in astronomy, NumPy was an important part of the software stack used in the discovery of gravitational waves <sup>1</sup> and in the first imaging of a black hole <sup>2</sup> . Here we review how a few fundamental array concepts lead to a simple and powerful programming paradigm for organizing, exploring and analysing scientific data. NumPy is the foundation upon which the scientific Python ecosystem is constructed. It is so pervasive that several projects, targeting audiences with specialized needs, have developed their own NumPy-like interfaces and array objects. Owing to its central position in the ecosystem, NumPy increasingly acts as an interoperability layer between such array computation libraries and, together with its application programming interface (API), provides a flexible framework to support the next decade of scientific and industrial analysis. </p>},
[…]
