### 2026-05-20 c0b7ad4
docs(paper): Fix references
