+++ paper/paper.bib
@@ -1,29 +1,32 @@
 @misc{Lam2026,
    author = {Siu Kwan Lam and stuartarchibald and Antoine Pitrou and Mark Florisson and Graham Markall and Stan Seibert and Emergency Self-Construct and Todd A Anderson and Guilherme Leobas and rjenc29 and Kaustubh and Swapnil Patel and Michael Collison and luk-f-a and Jay Bourque and Aaron Meurer and Travis E Oliphant and Michael Wang and densmirn and KrisMinchev and Andre Masella and Eric Wieser and njwhite and Ethan Pronovost and Ehsan Totoni and Stefan Seefeld and Hernan Grecco and Pearu Peterson and Isaac Virshup and MattyG},
-   doi = {10.5281/zenodo.19720971},
+   doi = {10.5281/zenodo.4343230},
    month = {4},
    publisher = {Zenodo},
    title = {numba},
-   url = {https://doi.org/10.5281/zenodo.19720971},
+   url = {https://doi.org/10.5281/zenodo.4343230},
    year = {2026}
 }
 @misc{Baratta2023,
    author = {Igor A Baratta and Joseph P Dean and Jørgen S Dokken and Michal Habera and Jack S Hale and Chris N Richardson and Marie E Rognes and Matthew W Scroggs and Nathan Sime and Garth N Wells},
-   doi = {10.5281/zenodo.10447666},
+   doi = {10.5281/zenodo.10447665},
    month = {12},
    publisher = {Zenodo},
-   title = {DOLFINx: The next generation FEniCS problem
-solving environment
-},
-   url = {https://doi.org/10.5281/zenodo.10447666},
+   title = {DOLFINx: The next generation FEniCS problem solving environment},
+   url = {https://doi.org/10.5281/zenodo.10447665},
    year = {2023}
+}
+@misc{Abaqus,
+   author = {{Dassault Systèmes Simulia Corp.}},
+   title = {Abaqus Analysis User's Manual, Version 2026},
+   year = {2026}
 }
 
 @misc{Schlomer2024,
    author = {Nico Schlömer},
    doi = {10.5281/zenodo.1173115},
    title = {meshio: Tools for mesh files},
-   url = {https://github.com/nschloe/meshio},
+   url = {https://doi.org/10.5281/zenodo.1173115},
    year = {2024}
 }
 @article{Harris2006,
@@ -38,7 +41,7 @@ solving environment
    pages = {357-362},
    pmid = {32939066},
    publisher = {Nature Publishing Group},
-   title = {Array programming with NumPy},
+   title = {Array programming with {NumPy}},
    volume = {585},
    url = {https://www.nature.com/articles/s41586-020-2649-2},
    year = {2020}
@@ -46,16 +49,15 @@ solving environment
 @misc{Zhou2019,
    author = {Qingnan Zhou},
    publisher = {Software available for download at https://github.com/PyMesh/PyMesh},
-   title = {Pymesh—geometry processing library for python},
-   volume = {7},
+   title = {Pymesh—geometry processing library for {Python}},
    year = {2019}
 }
 @misc{Kloeckner2025,
    author = {Andreas Kloeckner and Lorenz Brun and Benyuan Liu and Simon Klemenc and Alex Fkikl and Christoph Gohlke and Ethan Coon and Geoffrey Oxberry and Jozef Veselý and Matt Wala and Matt Smith and Peter Potrowl and Aidan Kurtz},
-   doi = {10.5281/ZENODO.15048253},
+   doi = {10.5281/zenodo.7296459},
    publisher = {Zenodo},
    title = {MeshPy},
-   url = {https://zenodo.org/records/15048253},
[…]
