### 2026-05-28 80a7bdc
docs(paper): Update references, AI usage disclosure
