+++ paper/paper.bib
@@ -1,3 +1,55 @@
+@misc{Stanford2026,
+   author = {{Stanford University Computer Graphics Laboratory}},
+   title = {The {Stanford} 3D Scanning Repository: {Stanford} Bunny},
+   url = {https://graphics.stanford.edu/data/3Dscanrep/},
+   year = {2026}
+}
+@article{Turk1994,
+   abstract = {Range imaging offers an inexpensive and accurate means for digitizing the shape of three-dimensional objects. Because most objects self occlude, no single range image suffices to describe the entire object. We present a method for combining a collection of range images into a single polygonal mesh that completely describes an object to the extent that it is visible from the outside. The steps in our method are: 1) align the meshes with each other using a modified iterated closest-point algorithm, 2) zipper together adjacent meshes to form a continuous surface that correctly captures the topology of the object, and 3) compute local weighted averages of surface positions on all meshes to form a consensus surface geometry. Our system differs from previous approaches in that it is incre-mental; scans are acquired and combined one at a time. This approach allows us to acquire and combine large numbers of scans with minimal storage overhead. Our largest models contain up to 360,000 triangles. All the steps needed to digitize an object that requires up to 10 range scans can be performed using our system with five minutes of user interaction and a few hours of compute time. We show two models created using our method with range data from a commercial rangefinder that employs laser stripe technology.},
+   author = {Greg Turk and Marc Levoy},
+   city = {New York, New York, USA},
+   doi = {10.1145/192161.192241},
+   isbn = {0897916670},
+   booktitle = {Proceedings of the 21st annual conference on Computer graphics and interactive techniques  - SIGGRAPH '94},
+   keywords = {CR Categories: I35 [Computer Graphics]: Computational Geom-etry and Object Modelling Additional Key Words: Surface reconstruction,polygon mesh,range images,structured light range scanner,surface fitting},
+   pages = {311-318},
+   publisher = {ACM Press},
+   title = {Zippered polygon meshes from range images},
+   url = {http://portal.acm.org/citation.cfm?doid=192161.192241},
+   year = {1994}
+}
+@article{Bezanson2017,
+   abstract = {Bridging cultures that have often been distant, Julia combines expertise from the diverse helds of computer science and computational science to create a new approach to numerical computing. Julia is designed to be easy and fast and questions notions generally held to be "laws of nature" by practitioners of numerical computing: 1. High-level dynamic programs have to be slow. 2. One must prototype in one language and then rewrite in another language for speed or deployment. 3. There are parts of a system appropriate for the programmer, and other parts that are best left untouched as they have been built by the experts. We introduce the Julia programming language and its design-a dance between specialization and abstraction. Specialization allows for custom treatment. Multiple dispatch, a technique from computer science, picks the right algorithm for the right circumstance. Abstraction, which is what good computation is really about, recognizes what remains the same after differences are stripped away. Abstractions in mathematics are captured as code through another technique from computer science, generic programming. Julia shows that one can achieve machine performance without sacrificing human convenience.},
+   author = {Jeff Bezanson and Alan Edelman and Stefan Karpinski and Viral B. Shah},
+   doi = {10.1137/141000671},
+   issn = {0036-1445},
+   issue = {1},
+   journal = {SIAM Review},
[…]
