+++ paper/paper.md
@@ -46,7 +46,7 @@ Meshes are used for a variety of purposes, including simulations (e.g. finite el
 Mesh-based representations of geometries are essential in a wide variety of research applications, and as such, there is a need for robust, efficient, and easy-to-use software for creating, analyzing, and manipulating meshes.
 There are a variety of software packages for working with and generating meshes. 
 Some are general purpose, like CGAL [@cgal], VTK [@Schroeder2006], and Gmsh [@Geuzaine2009], while others are more focused on specific tasks, such as triangular or tetrahedral mesh generation (e.g. Triangle [@Shewchuk1996] and TetGen [@Si2015], respectively). 
-In Python, most meshing packages depend on (or are direct wrappers to) one or more of these libraries, such as PyVista [@Sullivan2019] (a Pythonic interface to VTK), pygalmesh (a Pythonic interface to CGAL) [@Schlomer2021], nanomesh (an image-based meshing workflow tool that utilizes Triangle and TetGen) [@Smeets2022b], MeshPy (which interfaces to Triangle and TetGen) [@Kloeckner2025], and PyMesh (which depends on CGAL, Triangle, TetGen, and others) [@Zhou2019]. 
+In Python [@vanRossum1995], most meshing packages depend on (or are direct wrappers to) one or more of these libraries, such as PyVista [@Sullivan2019] (a Pythonic interface to VTK), pygalmesh (a Pythonic interface to CGAL) [@Schlomer2021], nanomesh (an image-based meshing workflow tool that utilizes Triangle and TetGen) [@Smeets2022b], MeshPy (which interfaces to Triangle and TetGen) [@Kloeckner2025], and PyMesh (which depends on CGAL, Triangle, TetGen, and others) [@Zhou2019]. 
 While these interfaces are useful and provide access to powerful mesh generation tools, their reliance on external dependencies can make them less easy to use and limit code readability, making it more difficult to understand how the code works. 
 TriMesh [@trimesh] stands out as a capable, pure-Python library focused on triangular surface meshes, but it isn't intended for use with quadrilateral, mixed-element, or volumetric meshes. 
 Given the intended focus and/or design philosophies of these existing softwares, it was determined that building `mymesh`, rather than making contributions to existing software, was the best way to achieve a full-featured, accessible, and easy to use Python package for creating and working with meshes. 
@@ -61,7 +61,7 @@ By providing an easily usable interface to both high-level and low-level functio
 # Research impact statement
 
 `mymesh` was originally developed in support of research within the Skeletal Mechanobiology and Biomechanics Lab at Boston University. 
-It was used extensively in the scaffold design optimization research by @Josephson2024b and is currently being used within multiple labs and institutions (e.g. @Lim2026, @Mtchedlishvili2026), for various ongoing projects including vertebral modeling, hip fracture modeling [@Olowu2026], growth modeling of skeletal tissue, and analysis of objects and biological tissues imaged using micro-computed tomography (μCT). 
+It was used extensively in the scaffold design optimization research by @Josephson2024b and is currently being used within multiple labs and institutions (e.g. Lim, M., pers. comm., Mtchedlishvili, M., pers. comm.), for various ongoing projects including vertebral modeling, hip fracture modeling [@Olowu2026], growth modeling of skeletal tissue, and analysis of objects and biological tissues imaged using micro-computed tomography (μCT). 
 `mymesh` has proven useful in a variety of research applications, well beyond those that inspired its original development, and we expect it to remain a valuable tool in future research efforts.  
 
 # Software design
@@ -75,7 +75,7 @@ The `mesh` object facilitates conversion to the data structures of two other pop
[…]
