### 2026-05-29 8e1eb5b
doxs(paper): fix doi
