+++ paper/paper.bib
@@ -135,7 +135,7 @@
 }
 @misc{Schlomer2021,
    author = {Nico Schlömer},
-   doi = {https://doi.org/10.5281/zenodo.5564818},
+   doi = {10.5281/zenodo.5564818},
    publisher = {Zenodo},
    title = {pygalmesh: Python interface for CGAL's meshing tools},
    url = {https://doi.org/10.5281/zenodo.5564818},
