### 2026-05-29 d13b017
Merge branch 'paper' into develop
