+++ .github/workflows/draft-pdf.yml
@@ -2,11 +2,7 @@ name: Draft PDF
 on:
   push:
     branches:
-<<<<<<< HEAD
-        - paper
-=======
       - paper
->>>>>>> paper
     paths:
       - paper/**
       - .github/workflows/draft-pdf.yml
