### 2026-05-30 cd2b136
minor changes in paper.md for JOSS
