+++ paper/paper.md
@@ -1,5 +1,5 @@
 ---
-title: 'MyMesh: General purpose, implicit, and image-based meshing in python'
+title: 'MyMesh: General purpose, implicit, and image-based meshing in Python'
 tags:
   - Python
   - mesh
@@ -29,23 +29,23 @@ bibliography: paper.bib
 # Summary
 
 A mesh is a discrete representation that subdivides a geometry or computational domain into a collection of points (nodes) connected by simple shapes (elements).
-Meshes are used for a variety of purposes, including simulations (e.g. finite element, finite volume, and finite difference methods), visualization and computer graphics, image analysis, and additive manufacturing.
+Meshes are used for a variety of purposes, including simulations (e.g., finite element, finite volume, and finite difference methods), visualization and computer graphics, image analysis, and additive manufacturing.
 `mymesh` is a general purpose set of tools for generating, manipulating, and analyzing meshes. 
 `mymesh` is particularly focused on implicit function and image-based meshing, with other functionality including:
 
 - geometric and curvature analysis,
-- intersection and inclusion tests (e.g. ray-surface intersection and point-in-surface tests),
+- intersection and inclusion tests (e.g., ray-surface intersection and point-in-surface tests),
 - mesh boolean operations (intersection, union, difference),
 - sweep construction methods (extrusions, revolutions),
 - point set, mesh, and image registration,
-- mesh quality evaluation and improvement,
+- mesh quality evaluation and improvement, and
 - mesh type conversion (e.g. volume to surface, hexahedral or mixed-element to tetrahedral, first-order elements to second-order elements).
 
 # State of the field
 
 Mesh-based representations of geometries are essential in a wide variety of research applications, and as such, there is a need for robust, efficient, and easy-to-use software for creating, analyzing, and manipulating meshes.
 There are a variety of software packages for working with and generating meshes. 
-Some are general purpose, like CGAL [@cgal], VTK [@Schroeder2006], and Gmsh [@Geuzaine2009], while others are more focused on specific tasks, such as triangular or tetrahedral mesh generation (e.g. Triangle [@Shewchuk1996] and TetGen [@Si2015], respectively). 
+Some are general purpose, like CGAL [@cgal], VTK [@Schroeder2006], and Gmsh [@Geuzaine2009], while others are more focused on specific tasks, such as triangular or tetrahedral mesh generation (e.g., Triangle [@Shewchuk1996] and TetGen [@Si2015], respectively). 
 In Python [@vanRossum1995], most meshing packages depend on (or are direct wrappers to) one or more of these libraries, such as PyVista [@Sullivan2019] (a Pythonic interface to VTK), pygalmesh (a Pythonic interface to CGAL) [@Schlomer2021], nanomesh (an image-based meshing workflow tool that utilizes Triangle and TetGen) [@Smeets2022b], MeshPy (which interfaces to Triangle and TetGen) [@Kloeckner2025], and PyMesh (which depends on CGAL, Triangle, TetGen, and others) [@Zhou2019]. 
 While these interfaces are useful and provide access to powerful mesh generation tools, their reliance on external dependencies can make them less easy to use and limit code readability, making it more difficult to understand how the code works. 
 TriMesh [@trimesh] stands out as a capable, pure-Python library focused on triangular surface meshes, but it isn't intended for use with quadrilateral, mixed-element, or volumetric meshes. 
@@ -61,13 +61,13 @@ By providing an easily usable interface to both high-level and low-level functio
 # Research impact statement
 
 `mymesh` was originally developed in support of research within the Skeletal Mechanobiology and Biomechanics Lab at Boston University. 
[…]
