### 2026-05-30 #10003 @danielskatz
@TimJosephson - As track editor, I've checked this submission.

In the zenodo archive, please edit the metadata to make python Python in the title.  Additionally, please merge the changes in https://github.com/BU-SMBL/mymesh/pull/10 or let me know what you disagree with.  I'm attempting to merge these into main, as it seems that that's the right place for the latest version of the paper, but I'm unsure - something odd seems to be going on.  If that's correct, please let me know and I'll make sure to build the final version from there. If not, please make the equivalent changes in paper, and I'll build the final version from there.
