### 2026-05-31 #10003 @danielskatz
@TimJosephson - please check the proof that this will hopefully generate to make sure everything is ok
