### 2026-06-02 337ce7b
docs: updates following JOSS publication
