+++ docs/source/dev/CHANGELOG.md
@@ -1,7 +1,11 @@
 <!-- Generated by the commitizen `cz changelog` command. See pyproject.toml for configuration -->
 # Change Log
 
-## Unreleased
+## 0.4.1 (2026-05-29)
+
+- Finalization of paper for [JOSS](https://joss.theoj.org/papers/10.21105/joss.10003)
+
+## 0.4.0 (2026-04-30)
 
 ### Feat
 
