+++ dev/_downloads/f4c392f9ffec2d542f031db250d48e22/tree-2.py
@@ -78,7 +78,7 @@ plt.loglog(npoints, np.add(kdtree_build_time, kdtree_search_time), color='#bf616
 
 plt.loglog(npoints, sci_kdtree_build_time, color='#a3be8c', linestyle='dashed')
 plt.loglog(npoints, sci_kdtree_search_time, color='#a3be8c', linestyle='dotted')
-plt.loglog(npoints, np.add(sci_kdtree_build_time, sci_kdtree_search_time), color='#a3be8c,  linestyle='solid')
+plt.loglog(npoints, np.add(sci_kdtree_build_time, sci_kdtree_search_time), color='#a3be8c',  linestyle='solid')
 
 
 plt.loglog(npoints, brute_search_time, color='black', linewidth=3)
