### 2026-08-09 b87537c
fix(PadRagged): add quick exit for inputs that are already arrays
