### 2026-08-09 4da0f22
fix(primitives.Grid): change NodeConn dtype to intp for faster downstream indexing
