### 2026-08-09 7361200
fix(contour.MarchingTetrahedra): efficiency enhancements
