+++ src/mymesh/contour.py
@@ -4676,6 +4676,25 @@ def MarchingTetrahedra(TetNodeCoords, TetNodeConn, NodeValues, threshold=0, inte
             []                      # 15-1111
         ],dtype=object)
 
+    MT_RLookup = np.array([
+                [-1,-1,-1,-1,-1,-1],    # 0-0000
+                [ 3, 5, 4,-1,-1,-1],    # 1-0001
+                [ 4, 2, 1,-1,-1,-1],    # 2-0010
+                [ 1, 3, 5, 1, 5, 2],    # 3-0011
+                [ 0, 3, 1,-1,-1,-1],    # 4-0100
+                [ 1, 0, 5, 1, 5, 4],    # 5-0101
+                [ 0, 3, 4, 0, 4, 2],    # 6-0110
+                [ 0, 5, 2,-1,-1,-1],    # 7-0111
+                [ 0, 2, 5,-1,-1,-1],    # 8-1000
+                [ 0, 2, 4, 0, 4, 3],    # 9-1001
+                [ 0, 1, 4, 0, 4, 5],    # 10-1010
+                [ 0, 1, 3,-1,-1,-1],    # 11-1011
+                [ 1, 2, 5, 1, 5, 3],    # 12-1100
+                [ 1, 2, 4,-1,-1,-1],    # 13-1101
+                [ 3, 4, 5,-1,-1,-1],    # 14-1110
+                [-1,-1,-1,-1,-1,-1]     # 15-1111
+            ])
+
     MTMixed_Lookup = np.array([
             [],             # 0-0000
             [[3, 5, 4]],    # 1-0001
@@ -4694,7 +4713,25 @@ def MarchingTetrahedra(TetNodeCoords, TetNodeConn, NodeValues, threshold=0, inte
             [[3, 4, 5]],    # 14-1110
             []              # 15-1111
         ],dtype=object)
-
+    MTMixed_RLookup = np.array([
+            [-1,-1,-1,-1],             # 0-0000
+            [ 3, 5, 4,-1],    # 1-0001
+            [ 4, 2, 1,-1],    # 2-0010
+            [ 1, 3, 5, 2], # 3-0011
+            [ 0, 3, 1,-1],    # 4-0100
+            [ 1, 0, 5, 4], # 5-0101
+            [ 0, 3, 4, 2], # 6-0110
+            [ 0, 5, 2,-1],    # 7-0111
+            [ 0, 2, 5,-1],    # 8-1000
+            [ 0, 2, 4, 3], # 9-1001
+            [ 0, 1, 4, 5], # 10-1010
+            [ 0, 1, 3,-1],    # 11-1011
+            [ 1, 2, 5, 3], # 12-1100
+            [ 1, 2, 4,-1],    # 13-1101
+            [ 3, 4, 5,-1],    # 14-1110
+            [-1,-1,-1,-1]              # 15-1111
+        ])
+    
     MTVMixed_Lookup = np.array([
             [],                     # 0-0000
             [[9, 3, 5, 4]],         # 1-0001
@@ -4713,6 +4750,24 @@ def MarchingTetrahedra(TetNodeCoords, TetNodeConn, NodeValues, threshold=0, inte
             [[7, 8, 6, 3, 4, 5]],   # 14-1110
             [[6, 7, 8, 9]]          # 15-1111
         ],dtype=object)
+    MTVMixed_RLookup = np.array([
+            [-1,-1,-1,-1,-1,-1],                     # 0-0000
+            [ 9, 3, 5, 4,-1,-1],         # 1-0001
+            [ 8, 4, 2, 1,-1,-1],         # 2-0010
[…]
