### 2026-08-09 f615c04
fix(contour.MarchingCubes): speedup marching cubes
