+++ src/mymesh/contour.py
@@ -283,6 +283,135 @@ MC_Lookup = np.array([
             [[10, 9, 6], [3, 0, 4]],
             [[4, 3, 0], [7, 11, 10]],
             [[4, 3, 0]]],dtype=object)
+MC_RLookup = np.array([
+    [-1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1, -1],
+    [ 7, 10, 11, -1, -1, -1, -1, -1, -1, -1, -1, -1],
+    [ 9, 10,  6, -1, -1, -1, -1, -1, -1, -1, -1, -1],
+    [ 6,  9,  7,  7,  9, 11, -1, -1, -1, -1, -1, -1],
+    [ 8,  9,  5, -1, -1, -1, -1, -1, -1, -1, -1, -1],
+    [ 5,  8,  9,  7, 10, 11, -1, -1, -1, -1, -1, -1],
+    [ 5,  8,  6,  6,  8, 10, -1, -1, -1, -1, -1, -1],
+    [ 5,  7,  6,  5,  8, 11,  5, 11,  7, -1, -1, -1],
+    [ 8,  4, 11, -1, -1, -1, -1, -1, -1, -1, -1, -1],
+    [ 7, 10,  4,  4, 10,  8, -1, -1, -1, -1, -1, -1],
+    [ 4, 11,  8,  6,  9, 10, -1, -1, -1, -1, -1, -1],
+    [ 7,  6,  4,  9,  8,  4,  6,  9,  4, -1, -1, -1],
+    [ 4,  9,  5, 11,  9,  4, -1, -1, -1, -1, -1, -1],
+    [ 7,  5,  4,  7, 10,  9,  7,  9,  5, -1, -1, -1],
+    [ 5,  4,  6, 11, 10,  6,  4, 11,  6, -1, -1, -1],
+    [ 4,  6,  5,  4,  7,  6, -1, -1, -1, -1, -1, -1],
+    [ 2,  7,  3, -1, -1, -1, -1, -1, -1, -1, -1, -1],
+    [10, 11,  2,  2, 11,  3, -1, -1, -1, -1, -1, -1],
+    [ 9, 10,  6,  3,  2,  7, -1, -1, -1, -1, -1, -1],
+    [ 3,  9, 11,  3,  2,  6,  3,  6,  9, -1, -1, -1],
+    [ 2,  7,  3,  5,  8,  9, -1, -1, -1, -1, -1, -1],
+    [ 5,  8,  9, 10,  3,  2, 11,  3, 10, -1, -1, -1],
+    [ 2,  7,  3,  5,  8,  6,  6,  8, 10, -1, -1, -1],
+    [ 3,  2,  6,  3,  6,  8, 11,  3,  8,  6,  5,  8],
+    [ 8,  4, 11,  2,  7,  3, -1, -1, -1, -1, -1, -1],
+    [ 8,  2, 10,  8,  4,  3,  8,  3,  2, -1, -1, -1],
+    [ 2,  7,  3,  9, 10,  6, 11,  8,  4, -1, -1, -1],
+    [ 2,  6,  3,  3,  6,  4,  6,  9,  4,  4,  9,  8],
+    [ 7,  3,  2,  9,  5, 11, 11,  5,  4, -1, -1, -1],
+    [10,  9,  5, 10,  5,  3,  2, 10,  3,  5,  4,  3],
+    [ 2,  6,  3,  3,  6,  4,  6,  5,  4,  7, 11, 10],
+    [ 4,  6,  5,  4,  3,  2,  4,  2,  6, -1, -1, -1],
[…]
